### 2024-07-24 #7029 @crvernon
:wave:  @slinderman - thanks for the submission.  While I get an editor set up for you, could you please add a "Statement of Need" section to the paper and address the missing DOI in the message above if they are applicable.  Also, I appreciate you keeping the word count at around 1000 words - thank you!
