+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={249--275},
   year={2020},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi={https://doi.org/10.1146/annurev-neuro-092619-094115}
 }
 
  @book{murphy2023probabilistic,
@@ -22,7 +23,8 @@
   author={S{\"a}rkk{\"a}, Simo and Svensson, Lennart},
   volume={17},
   year={2023},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={https://doi.org/10.1017/CBO9781139344203}
 }
 
 
@@ -40,7 +42,8 @@
   booktitle={International Conference on Machine Learning},
   pages={42046--42057},
   year={2023},
-  organization={PMLR}
+  organization={PMLR},
+  doi={https://doi.org/10.48550/arXiv.2305.16543}
 }
 
 @article{lee2023switching,
@@ -49,7 +52,8 @@
   journal={Advances in Neural Information Processing Systems},
   volume={36},
   pages={57976--58010},
-  year={2023}
+  year={2023},
+  doi={https://doi.org/10.48550/arXiv.2306.03291}
 }
 
 @inproceedings{chang2023low,
@@ -63,6 +67,7 @@
   series = 	 {Proceedings of Machine Learning Research},
   month = 	 {22--25 Aug},
   publisher =    {PMLR},
+  doi={https://doi.org/10.48550/arXiv.2305.19535},
 }
 
 
@@ -77,6 +82,7 @@
   title = {Keypoint-{M}o{S}eq: parsing behavior by linking point tracking to pose dynamics},
   volume = {21},
   year = {2024},
+  doi={https://doi.org/10.1038/s41592-024-02318-2},
 }
 
 @misc{pyhsmm,
@@ -116,6 +122,7 @@ year={2022}
   author={Seabold, Skipper and Perktold, Josef},
   booktitle={9th Python in Science Conference},
   year={2010},
+  doi={https://doi.org/10.25080/majora-92bf1922-011}
 }
 
 @misc{hmmlearn,
@@ -133,7 +140,8 @@ year={2022}
                Graeme",
   publisher = "Cambridge University Press",
   month     =  apr,
-  year      =  1998
+  year      =  1998,
+  doi={https://doi.org/10.1017/cbo9780511790492},
 }
 
 @article{patterson2008state,
@@ -144,7 +152,8 @@ year={2022}
   number={2},
   pages={87--94},
   year={2008},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.tree.2007.10.009}
 }
 
 @article{jacquier2002bayesian,
@@ -155,7 +164,8 @@ year={2022}
   number={1},
   pages={69--87},
   year={2002},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={https://doi.org/10.1198/073500102753410408}
 }
 
 @article{ott2004local,
@@ -166,7 +176,8 @@ year={2022}
   number={5},
   pages={415--428},
   year={2004},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={https://doi.org/10.3402/tellusa.v56i5.14462}
 }
 
 @article{stone1975parallel,
@@ -177,7 +188,8 @@ year={2022}
   number={4},
   pages={289--307},
   year={1975},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={https://doi.org/10.1145/355656.355657}
 }
 
[…]
