+++ paper/paper.bib
@@ -7,7 +7,7 @@
   pages={249--275},
   year={2020},
   publisher={Annual Reviews},
-  doi={https://doi.org/10.1146/annurev-neuro-092619-094115}
+  doi={10.1146/annurev-neuro-092619-094115}
 }
 
  @book{murphy2023probabilistic,
@@ -24,7 +24,7 @@
   volume={17},
   year={2023},
   publisher={Cambridge University Press},
-  doi={https://doi.org/10.1017/CBO9781139344203}
+  doi={10.1017/CBO9781139344203}
 }
 
 
@@ -43,7 +43,7 @@
   pages={42046--42057},
   year={2023},
   organization={PMLR},
-  doi={https://doi.org/10.48550/arXiv.2305.16543}
+  doi={10.48550/arXiv.2305.16543}
 }
 
 @article{lee2023switching,
@@ -53,7 +53,7 @@
   volume={36},
   pages={57976--58010},
   year={2023},
-  doi={https://doi.org/10.48550/arXiv.2306.03291}
+  doi={10.48550/arXiv.2306.03291}
 }
 
 @inproceedings{chang2023low,
@@ -67,7 +67,7 @@
   series = 	 {Proceedings of Machine Learning Research},
   month = 	 {22--25 Aug},
   publisher =    {PMLR},
-  doi={https://doi.org/10.48550/arXiv.2305.19535},
+  doi={10.48550/arXiv.2305.19535},
 }
 
 
@@ -82,7 +82,7 @@
   title = {Keypoint-{M}o{S}eq: parsing behavior by linking point tracking to pose dynamics},
   volume = {21},
   year = {2024},
-  doi={https://doi.org/10.1038/s41592-024-02318-2},
+  doi={10.1038/s41592-024-02318-2},
 }
 
 @misc{pyhsmm,
@@ -122,7 +122,7 @@ year={2022}
   author={Seabold, Skipper and Perktold, Josef},
   booktitle={9th Python in Science Conference},
   year={2010},
-  doi={https://doi.org/10.25080/majora-92bf1922-011}
+  doi={10.25080/majora-92bf1922-011}
 }
 
 @misc{hmmlearn,
@@ -141,7 +141,7 @@ year={2022}
   publisher = "Cambridge University Press",
   month     =  apr,
   year      =  1998,
-  doi={https://doi.org/10.1017/cbo9780511790492},
+  doi={10.1017/cbo9780511790492},
 }
 
 @article{patterson2008state,
@@ -153,7 +153,7 @@ year={2022}
   pages={87--94},
   year={2008},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.tree.2007.10.009}
+  doi={10.1016/j.tree.2007.10.009}
 }
 
 @article{jacquier2002bayesian,
@@ -165,7 +165,7 @@ year={2022}
   pages={69--87},
   year={2002},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.1198/073500102753410408}
+  doi={10.1198/073500102753410408}
 }
 
 @article{ott2004local,
@@ -177,7 +177,7 @@ year={2022}
   pages={415--428},
   year={2004},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.3402/tellusa.v56i5.14462}
+  doi={10.3402/tellusa.v56i5.14462}
 }
 
 @article{stone1975parallel,
@@ -189,7 +189,7 @@ year={2022}
   pages={289--307},
   year={1975},
   publisher={ACM New York, NY, USA},
-  doi={https://doi.org/10.1145/355656.355657}
+  doi={10.1145/355656.355657}
 }
 
[…]
