### 2024-07-24 #7029 @slinderman
Thanks, @crvernon! I've added the missing DOIs. Please note that some online resources (e.g., GitHub repositories for related packages) do not have DOIs. I also confirmed that the paper has a "Statement of need" section.
