### 2024-07-24 #7029 @osorensen
👋 @Saran-nns, @GarrettMerz, @edeno, @thomaspinder, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
