### 2024-08-03 #7029 @osorensen
👋 @gdalle, @florisvb, @bvdmitri, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
