### 2024-08-04 #7069 @editorialbot
[REVIEW]: Dynamax: A Python package for probabilistic state space modeling with JAX
**Submitting author:** @slinderman (<a href="http://orcid.org/0000-0002-3878-9073">Scott Linderman</a>)
**Repository:** https://github.com/probml/dynamax
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.0
