### 2024-08-12 #7029 @florisvb
Sorry I was out of town the last 2 weeks, but it looks like you have enough
reviewers for this now?

- Floris

On Sat, Aug 3, 2024 at 2:22 AM Øystein Sørensen ***@***.***>
wrote:

> 👋 @gdalle <https://github.com/gdalle>, @florisvb
> <https://github.com/florisvb>, @bvdmitri <https://github.com/bvdmitri>,
> would any of you be willing to review this submission for JOSS? We carry
> out our checklist-driven reviews here in GitHub issues and follow these
> guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
