### 2024-09-01 e7a35f8
Add interrogate workflow
