+++ .github/workflows/interrogate.yml
@@ -1,0 +1,28 @@
+name: Docstrings
+
+on:
+  push:
+    branches:
+    - main
+  pull_request:
+    branches:
+    - main
+
+jobs:
+  Workflow:
+    runs-on: ubuntu-latest
+    steps:
+    - name: Clone the reference repository
+      uses: actions/checkout@v
+
+    - name: Set up Python 3.10.6
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.10.6'
+
+    - name: Install dependencies
+      run: |
+        pip install -e '.[dev]'
+
+    - name: Run docstring checks
+      run: interrogate -vv dynamax --fail-under 66
