+++ .github/workflows/interrogate.yml
@@ -13,10 +13,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - name: Clone the reference repository
-      uses: actions/checkout@v
+      uses: actions/checkout@v3.5.2
 
     - name: Set up Python 3.10.6
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v4
       with:
         python-version: '3.10.6'
 
