### 2024-09-01 #7069 @thomaspinder
# Dynamax Review

Firstly, I'd like to congratulate the authors and contributors of Dynamax for creating a nicely designed package in JAX. Dynamax enables practitioners to easily fit state-space models (SSM), whilst simultaneously allowing researchers to implement their own custom SSM approaches. The repo's corresponding paper is well written and provides a clear and concise summary of the paper. Finally, I enjoyed reading the documentation of Dynamax; the large number of examples and use-cases is great for new users of the package.

I have broken my review up into two sections. The first section lists the blocking issues I have that prevent me from being able to mark each item of the reviewer's checklist as complete. In the second section, I suggest some things that I would advise the authors to do in Dynamax; however, I do not consider these as blocking concerns.

## Significant issues

- [ ] Failing tests
- [ ] Docstring coverage
- [ ] Incomplete contribution guidelines
- [ ] Missing typing

### Failing tests

For me, running
```
git clone git@github.com:probml/dynamax.git
cd dynamax
pip install -e '.[dev]'
pytest dynamax
```
as per the docs, threw errors on a Mac M1. I was using Python 3.10 and a fresh virtual environment.

On this topic though, I see that you're Github testing workflow uses a pinned Python version and machine. It could be good to run your tests on all supported Python versions and a Mac and Linux machine. I have opened a PR with a suggestion that reflects this comment.

### Docstring coverage

The docstrings within Dynamax are inconsistent. Taking, for example, `dynamax.generalized_gaussian_ssm.inference` as an example file, there are no docstrings for the methods or arguments of the public classes, yet there are comprehensive docstrings for the private functions. This is unconventional, and I would encourage the authors to more widely document the classes, methods, and functions of Dynamax so as to make the package easier for people to use. Additionally, typing such as the below from `dynamax.generalized_gaussian_ssm.models` is not particularly descriptive.

```python
:param initial_mean: $m$
:param initial_covariance: $S$
```
It would be good to see more comprehensive documentation. To rigorously test documentation thresholds in the future, you may consider using `interrogate`. This will measure the number of objects for which a docstring is supplied and raise errors/warnings when the coverage decreases. I have opened a suggestive PR for how this can be achieved.

33.1% of classes/methods/functions are covered by tests. It would be good to see this increased. This was calculated with `interrogate dynamax`.

### Broken documentation

The documentation is broken and many notebooks do not render - [Example](https://probml.github.io/dynamax/notebooks/linear_gaussian_ssm/kf_linreg.html). My suggestion would be to use something like nbsphinx to execute the notebooks each time the documentation is built, and to fail when a notebook does not execute top-to-bottom.

### Incomplete contribution guidelines

Your [contribution guidelines](https://github.com/probml/dynamax/blob/main/CONTRIBUTING.md) are incomplete. To push a change, one would need to add and commit the change before pushing - this is missing from the document.

### Missing typing

The package has no typing. Additionally, many functions use variable names which are very hard to interpret e.g., `def _predict(m, P, f, Q, u, g_ev, g_cov)` in `dynamax.generalized_gaussian_ssm.inference`. The combination of no typing and terse variable naming makes the code challenging to read, particularly for new users and/or those new to SSMs. I would strongly encourage the authors to either use more descriptive variable names, or add typing to their code. JaxTyping is excellent for this.

### Incorrect typing

Some of the typing used is incorrectly done. Take `ParamsGGSSM.initial_mean` in `dynamax.generalized_gaussian_ssm.models` as an example. In this function, the typing on `initial_mean` is incorrect, as [outlined here in JaxTyping](https://docs.kidger.site/jaxtyping/faq/#flake8-or-ruff-are-throwing-an-error). A change to the effect of
```diff
- initial_mean: Float[Array, "state_dim"]
+ initial_mean: Float[Array, " state_dim"]
```
should be made.

## Suggested improvements

- [ ] Large repo size
- [ ] Tighter Python bounds
- [ ] Package management

### Large repo size

You have some very large git pack files:
```
❯ du -ah dynamax | sort -rh | head -n 10

224M	dynamax
209M	dynamax/.git/objects/pack
209M	dynamax/.git/objects
209M	dynamax/.git
208M	dynamax/.git/objects/pack/pack-aaa194639683ed252462b75ee0ec2203edc5c09f.pack
```
You may consider running `git gc` to clean these up unless you need them.

### Tighter Python bounds

I would suggest tighter Python bounds. Currently your lower bound is 3.6; a version which is considered end-of-life by the Python Org.: https://devguide.python.org/versions/#unsupported-versions
As an aside, I appreciate that it's specified in your setup.py, but explicitly stating the supported Python version(s) in your project's README is nice.

### Package management

It is [strongly advised](https://packaging.python.org/en/latest/guides/modernize-setup-py-project/#should-pyproject-toml-be-added) to migrate `setup.py` and `setup.cfg` files to a `pyproject.toml`. I would encourage the authors to do this in order to future-proof Dynamax.
