### 2024-09-11 #7069 @slinderman
Thank you, @osorensen and @thomaspinder! We will start working on these issues and suggestions ASAP, and I'll keep you posted.
