### 2024-09-12 probml/dynamax#377 @gdalle
JOSS review - Remarks on code
Hi and congrats on the package!

I'm one of the reviewers for the JOSS paper you submitted, so here I'll list my questions and concerns about the code itself. This issue will be updated as my reading progresses so maybe don't start answering right away.

> [!WARNING]
> I am not a Python developer, my main language is Julia, so I won't be able to give much advice beyond what I observed when trying out the examples. In addition, I don't know the best practices in terms of package and version management, I just use mamba and then pip inside the mamba environment.

---

## Execution

**Tests**

I followed the testing workflow on the home page of the documentation and it yielded one failure.

```python
============================== test session starts ===============================
rootdir: /home/gdalle/Work/Review/dynamax
configfile: pyproject.toml
plugins: cov-5.0.0, jaxtyping-0.2.34, typeguard-2.13.3, anyio-4.4.0
collected 83 items

dynamax/generalized_gaussian_ssm/inference_test.py .                                                                                                                  [  1%]
dynamax/generalized_gaussian_ssm/models_test.py ...                                                                                                                   [  4%]
dynamax/hidden_markov_model/inference_test.py ...........                                                                                                             [ 18%]
dynamax/hidden_markov_model/models/test_models.py ...........F......                                                                     [ 39%]
dynamax/linear_gaussian_ssm/inference_test.py ...                                                                                        [ 43%]
dynamax/linear_gaussian_ssm/info_inference_test.py .......                                                                               [ 51%]
dynamax/linear_gaussian_ssm/models_test.py ..                                                                                            [ 54%]
dynamax/linear_gaussian_ssm/parallel_inference_test.py ............                                                                      [ 68%]
dynamax/nonlinear_gaussian_ssm/inference_ekf_test.py .....                                                                               [ 74%]
dynamax/nonlinear_gaussian_ssm/inference_ukf_test.py .                                                                                   [ 75%]
dynamax/parameters_test.py ....                                                                                                          [ 80%]
dynamax/slds/inference_test.py ..                                                                                                        [ 83%]
dynamax/utils/distributions_test.py ..........                                                                                           [ 95%]
dynamax/utils/utils_test.py ....                                                                                                         [100%]

=================================================================== FAILURES ===================================================================
__________________________________________ test_sample_and_fit[LinearRegressionHMM-kwargs11-inputs11] __________________________________________

cls = <class 'dynamax.hidden_markov_model.models.linreg_hmm.LinearRegressionHMM'>, kwargs = {'emission_dim': 3, 'input_dim': 5, 'num_states': 4}
inputs = Array([[1., 1., 1., 1., 1.],
       [1., 1., 1., 1., 1.],
       [1., 1., 1., 1., 1.],
       [1., 1., 1., 1., 1.],
  ...1., 1., 1., 1.],
       [1., 1., 1., 1., 1.],
       [1., 1., 1., 1., 1.],
       [1., 1., 1., 1., 1.]], dtype=float32)

    @pytest.mark.parametrize(["cls", "kwargs", "inputs"], CONFIGS)
    def test_sample_and_fit(cls, kwargs, inputs):
        hmm = cls(**kwargs)
        #key1, key2 = jr.split(jr.PRNGKey(int(datetime.now().timestamp())))
        key1, key2 = jr.split(jr.PRNGKey(42))
        params, param_props = hmm.initialize(key1)
        states, emissions = hmm.sample(params, key2, num_timesteps=NUM_TIMESTEPS, inputs=inputs)
        fitted_params, lps = hmm.fit_em(params, param_props, emissions, inputs=inputs, num_iters=10)
>       assert monotonically_increasing(lps, atol=1e-2, rtol=1e-2)
E       assert Array(False, dtype=bool)
E        +  where Array(False, dtype=bool) = monotonically_increasing(Array([-224.90086,        nan,        nan,        nan,        nan,\n              nan,        nan,        nan,        nan,        nan],      dtype=float32), atol=0.01, rtol=0.01)

dynamax/hidden_markov_model/models/test_models.py:39: AssertionError
------------------------------------------------------------- Captured stdout call -------------------------------------------------------------
 |████████████████████████████████████████| 100.00% [10/10 00:00<00:00]
=============================================================== warnings summary ===============================================================
dynamax/generalized_gaussian_ssm/inference.py:86
  /home/gdalle/Work/Review/dynamax/dynamax/generalized_gaussian_ssm/inference.py:86: SyntaxWarning: invalid escape sequence '\i'
    """Predict next mean and covariance under an additive-noise Gaussian filter

dynamax/hidden_markov_model/models/abstractions.py:488
  /home/gdalle/Work/Review/dynamax/dynamax/hidden_markov_model/models/abstractions.py:488: SyntaxWarning: invalid escape sequence '\m'
    """Abstract base class of Hidden Markov Models (HMMs).

dynamax/hidden_markov_model/models/gaussian_hmm.py:722
  /home/gdalle/Work/Review/dynamax/dynamax/hidden_markov_model/models/gaussian_hmm.py:722: SyntaxWarning: invalid escape sequence '\s'
    """Initialize the model parameters and their corresponding properties.

dynamax/hidden_markov_model/parallel_inference.py:15
  /home/gdalle/Work/Review/dynamax/dynamax/hidden_markov_model/parallel_inference.py:15: SyntaxWarning: invalid escape sequence '\m'
    """Filtering associative scan elements.

dynamax/hidden_markov_model/parallel_inference.py:135
  /home/gdalle/Work/Review/dynamax/dynamax/hidden_markov_model/parallel_inference.py:135: SyntaxWarning: invalid escape sequence '\m'
    """Associative scan elements $E_ij$ are vectors specifying a sample::

dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key0-kwargs0]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key0-kwargs0]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key1-kwargs1]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key1-kwargs1]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key2-kwargs2]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key2-kwargs2]
dynamax/hidden_markov_model/models/test_models.py::test_sample_and_fit[PoissonHMM-kwargs14-None]
dynamax/hidden_markov_model/models/test_models.py::test_sample_and_fit[PoissonHMM-kwargs14-None]
  /home/gdalle/miniforge3/envs/dynamax-test/lib/python3.12/site-packages/jax/_src/numpy/array_methods.py:118: UserWarning: Explicitly requested dtype <class 'jax.numpy.float64'> requested in astype is not available, and will be truncated to dtype float32. To enable more dtypes, set the jax_enable_x64 configuration option or the JAX_ENABLE_X64 shell environment variable. See https://github.com/google/jax#current-gotchas for more.
    return lax_numpy.astype(self, dtype, copy=copy, device=device)

dynamax/generalized_gaussian_ssm/models_test.py: 12 warnings
dynamax/hidden_markov_model/models/test_models.py: 4 warnings
  /home/gdalle/miniforge3/envs/dynamax-test/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/ops.py:339: UserWarning: Explicitly requested dtype <class 'jax.numpy.float64'> requested in array is not available, and will be truncated to dtype float32. To enable more dtypes, set the jax_enable_x64 configuration option or the JAX_ENABLE_X64 shell environment variable. See https://github.com/google/jax#current-gotchas for more.
    return np.array(value, dtype=dtype)

dynamax/generalized_gaussian_ssm/models_test.py: 12 warnings
dynamax/hidden_markov_model/models/test_models.py: 4 warnings
  /home/gdalle/miniforge3/envs/dynamax-test/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/random_generators.py:290: UserWarning: Explicitly requested dtype <class 'jax.numpy.float64'> requested in zeros is not available, and will be truncated to dtype float32. To enable more dtypes, set the jax_enable_x64 configuration option or the JAX_ENABLE_X64 shell environment variable. See https://github.com/google/jax#current-gotchas for more.
    minval = minval + np.zeros([1] * final_rank, dtype=dtype)

dynamax/generalized_gaussian_ssm/models_test.py: 12 warnings
dynamax/hidden_markov_model/models/test_models.py: 4 warnings
  /home/gdalle/miniforge3/envs/dynamax-test/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/random_generators.py:291: UserWarning: Explicitly requested dtype <class 'jax.numpy.float64'> requested in zeros is not available, and will be truncated to dtype float32. To enable more dtypes, set the jax_enable_x64 configuration option or the JAX_ENABLE_X64 shell environment variable. See https://github.com/google/jax#current-gotchas for more.
    maxval = maxval + np.zeros([1] * final_rank, dtype=dtype)

dynamax/generalized_gaussian_ssm/models_test.py: 12 warnings
dynamax/hidden_markov_model/models/test_models.py: 4 warnings
  /home/gdalle/miniforge3/envs/dynamax-test/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/random_generators.py:292: UserWarning: Explicitly requested dtype <class 'jax.numpy.float64'>  is not available, and will be truncated to dtype float32. To enable more dtypes, set the jax_enable_x64 configuration option or the JAX_ENABLE_X64 shell environment variable. See https://github.com/google/jax#current-gotchas for more.
    return jaxrand.uniform(key=seed, shape=shape, dtype=dtype, minval=minval,

dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key0-kwargs0]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key0-kwargs0]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key1-kwargs1]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key1-kwargs1]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key2-kwargs2]
dynamax/generalized_gaussian_ssm/models_test.py::test_poisson_emission[key2-kwargs2]
dynamax/hidden_markov_model/models/test_models.py::test_sample_and_fit[PoissonHMM-kwargs14-None]
dynamax/hidden_markov_model/models/test_models.py::test_sample_and_fit[PoissonHMM-kwargs14-None]
  /home/gdalle/miniforge3/envs/dynamax-test/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/numpy_array.py:450: UserWarning: Explicitly requested dtype <class 'jax.numpy.float64'> requested in ones is not available, and will be truncated to dtype float32. To enable more dtypes, set the jax_enable_x64 configuration option or the JAX_ENABLE_X64 shell environment variable. See https://github.com/google/jax#current-gotchas for more.
    lambda shape, dtype=np.float32, name=None, layout=None: np.ones(  # pylint: disable=g-long-lambda

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
=========================================================== short test summary info ============================================================
FAILED dynamax/hidden_markov_model/models/test_models.py::test_sample_and_fit[LinearRegressionHMM-kwargs11-inputs11] - assert Array(False, dtype=bool)
============================================ 1 failed, 82 passed, 85 warnings in 149.15s (0:02:29) =============================================
```

**First tutorial**

When I try to run the [tutorial notebook](https://github.com/probml/dynamax/blob/main/docs/notebooks/hmm/gaussian_hmm.ipynb) after naive installation, it fails to generate sample data due to a Numpy 2.0 breaking change. I'm not sure why pip allowed me to install Numpy 2.0 as a dependency, since it doesn't seem to be supported yet (#366).

```python
Cell In[5], [line 22](vscode-notebook-cell:?execution_count=5&line=22)
     [15](vscode-notebook-cell:?execution_count=5&line=15) emission_means = jnp.column_stack([
     [16](vscode-notebook-cell:?execution_count=5&line=16)     jnp.cos(jnp.linspace(0, 2 * jnp.pi, true_num_states + 1))[:-1],
     [17](vscode-notebook-cell:?execution_count=5&line=17)     jnp.sin(jnp.linspace(0, 2 * jnp.pi, true_num_states + 1))[:-1],
     [18](vscode-notebook-cell:?execution_count=5&line=18)     jnp.zeros((true_num_states, emission_dim - 2)),
     [19](vscode-notebook-cell:?execution_count=5&line=19)     ])
     [20](vscode-notebook-cell:?execution_count=5&line=20) emission_covs = jnp.tile(0.1**2 * jnp.eye(emission_dim), (true_num_states, 1, 1))
---> [22](vscode-notebook-cell:?execution_count=5&line=22) true_params, _ = hmm.initialize(initial_probs=initial_probs,
     [23](vscode-notebook-cell:?execution_count=5&line=23)                                 transition_matrix=transition_matrix,
     [24](vscode-notebook-cell:?execution_count=5&line=24)                                 emission_means=emission_means,
     [25](vscode-notebook-cell:?execution_count=5&line=25)                                 emission_covariances=emission_covs)
     [27](vscode-notebook-cell:?execution_count=5&line=27) # Sample train, validation, and test data
     [28](vscode-notebook-cell:?execution_count=5&line=28) train_key, val_key, test_key = jr.split(jr.PRNGKey(0), 3)

File ~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:651, in GaussianHMM.initialize(self, key, method, initial_probs, transition_matrix, emission_means, emission_covariances, emissions)
    [649](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:649) params["initial"], props["initial"] = self.initial_component.initialize(key1, method=method, initial_probs=initial_probs)
    [650](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:650) params["transitions"], props["transitions"] = self.transition_component.initialize(key2, method=method, transition_matrix=transition_matrix)
--> [651](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:651) params["emissions"], props["emissions"] = self.emission_component.initialize(key3, method=method, emission_means=emission_means, emission_covariances=emission_covariances, emissions=emissions)
    [652](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:652) return ParamsGaussianHMM(**params), ParamsGaussianHMM(**props)

File ~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:83, in GaussianHMMEmissions.initialize(self, key, method, emission_means, emission_covariances, emissions)
     [81](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:81) elif method.lower() == "prior":
     [82](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:82)     this_key, key = jr.split(key)
---> [83](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:83)     prior = NormalInverseWishart(self.emission_prior_mean, self.emission_prior_conc,
     [84](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:84)                                  self.emission_prior_df, self.emission_prior_scale)
     [85](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:85)     (_emission_covs, _emission_means) = prior.sample(seed=this_key, sample_shape=(self.num_states,))
     [87](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py:87) else:

File ~/miniforge3/envs/dynamax/lib/python3.12/site-packages/decorator.py:232, in decorate.<locals>.fun(*args, **kw)
    [230](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/decorator.py:230) if not kwsyntax:
    [231](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/decorator.py:231)     args, kw = fix(args, kw, sig)
--> [232](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/decorator.py:232) return caller(func, *(extras + args), **kw)

File ~/miniforge3/envs/dynamax/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py:342, in _DistributionMeta.__new__.<locals>.wrapped_init(***failed resolving arguments***)
    [339](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py:339) # Note: if we ever want to have things set in `self` before `__init__` is
    [340](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py:340) # called, here is the place to do it.
    [341](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py:341) self_._parameters = None
--> [342](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py:342) default_init(self_, *args, **kwargs)
    [343](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py:343) # Note: if we ever want to override things set in `self` by subclass
    [344](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py:344) # `__init__`, here is the place to do it.
    [345](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py:345) if self_._parameters is None:
    [346](https://file+.vscode-resource.vscode-cdn.net/home/gdalle/Work/Review/~/miniforge3/envs/dynamax/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py:346)   # We prefer subclasses will set `parameters = dict(locals())` because
[…]
