### 2024-09-12 #7069 @gdalle
Meanwhile I'm making my way through my own review. I will gather my remarks about code and documentation in two issues on the dynamax repo:

- https://github.com/probml/dynamax/issues/376
- https://github.com/probml/dynamax/issues/377

I might be limited by my very ancient Python skills for installation and testing of dynamax, so I'd appreciate a hand in figuring out the errors I observe.

As for the paper itself, it is very clear and does a good job of introducing dynamax. I have three minor suggestions:

> Dynamax supports 39 canonical SSMs and allows the user to construct bespoke models as needed.

Can you give more details on how this is implemented API-wise? For instance, how generic can observation distributions be?
This question was a major motivation for my own take on HMMs, coded in Julia (<https://joss.theoj.org/papers/10.21105/joss.06436>). Hopefully I didn't misrepresent dynamax in the state of the art there.

> Dynamax provides a unique combination of low-level inference algorithms and high-level 53 modeling objects that can support a wide range of research applications in JAX.

Would it be possible to provide concrete examples of what is missing from the previous libraries? Obviously JAX support is a big aspect, since I know that some of them are coded in Numpy (hmmlearn) or PyTorch (pomegranate)

> Parallel message passing routines that leverage GPU or TPU acceleration to perform 63 message passing in sublinear time.

What do you mean by sublinear time? Isn't it just (roughly speaking) total sequential time divided by amount of parallelism?
