### 2024-09-12 df1c1d4
Change input and shape of LinRegHMM test to fix failure
Changes:
- Replace `jnp.ones` input with `jr.normal`.
- Reduce size of hidden state to 3.
- Remove unused `datetime` import and commented lines.

Fundamentally the problem is that the solve step can be unstable. This
does not resolve that but instead chooses a set up which is less
vulnerable to the instability.
