### 2024-09-12 20a1df7
Prepend space in uni-dim jaxtyping hints
Resolves F821
- see: https://docs.kidger.site/jaxtyping/faq/#flake8-or-ruff-are-throwing-an-error
