+++ dynamax/hidden_markov_model/models/categorical_glm_hmm.py
@@ -120,8 +120,8 @@ class CategoricalRegressionHMM(HMM):
                  num_states: int,
                  num_classes: int,
                  input_dim: int,
-                 initial_probs_concentration: Union[Scalar, Float[Array, "num_states"]]=1.1,
-                 transition_matrix_concentration: Union[Scalar, Float[Array, "num_states"]]=1.1,
+                 initial_probs_concentration: Union[Scalar, Float[Array, " num_states"]]=1.1,
+                 transition_matrix_concentration: Union[Scalar, Float[Array, " num_states"]]=1.1,
                  transition_matrix_stickiness: Scalar=0.0,
                  m_step_optimizer: optax.GradientTransformation=optax.adam(1e-2),
                  m_step_num_iters: int=50):
@@ -138,7 +138,7 @@ class CategoricalRegressionHMM(HMM):
     def initialize(self,
                    key: jr.PRNGKey=jr.PRNGKey(0),
                    method: str="prior",
-                   initial_probs: Optional[Float[Array, "num_states"]]=None,
+                   initial_probs: Optional[Float[Array, " num_states"]]=None,
                    transition_matrix: Optional[Float[Array, "num_states num_states"]]=None,
                    emission_weights: Optional[Float[Array, "num_states num_classes input_dim"]]=None,
                    emission_biases: Optional[Float[Array, "num_states num_classes"]]=None,
