+++ dynamax/slds/models.py
@@ -1,27 +1,15 @@
-from fastprogress.fastprogress import progress_bar
-from functools import partial
-from jax import jit, lax
+from jax import lax
 import jax.numpy as jnp
 import jax.random as jr
 from jax.tree_util import tree_map
-from jaxtyping import Array, Float, PyTree
+from jaxtyping import Array, Float
 import tensorflow_probability.substrates.jax.distributions as tfd
 from tensorflow_probability.substrates.jax.distributions import MultivariateNormalFullCovariance as MVN
-from typing import Any, Optional, Tuple, Union
-from typing_extensions import Protocol
+from typing import Optional, Tuple
 
 from dynamax.ssm import SSM
-from dynamax.linear_gaussian_ssm.models import LinearGaussianSSM
-from dynamax.linear_gaussian_ssm.inference import lgssm_filter, lgssm_smoother, lgssm_posterior_sample
 from dynamax.slds.inference import ParamsSLDS
-from dynamax.linear_gaussian_ssm.inference import PosteriorGSSMFiltered, PosteriorGSSMSmoothed
-from dynamax.parameters import ParameterProperties, ParameterSet
-from dynamax.types import PRNGKey, Scalar
-from dynamax.utils.bijectors import RealToPSDBijector
-from dynamax.utils.distributions import MatrixNormalInverseWishart as MNIW
-from dynamax.utils.distributions import NormalInverseWishart as NIW
-from dynamax.utils.distributions import mniw_posterior_update, niw_posterior_update
-from dynamax.utils.utils import pytree_stack, psd_solve
+from dynamax.types import PRNGKey
 
 class SLDS(SSM):
     
