### 2024-09-12 547c610
Fix LinearGaussianSSM.sample type hint
