### 2024-09-12 62df3ba
Update HMM[Parameter|Property]Set protocols
- Change attributes of these protocols to read-only.
  - This change means that instances of `NamedTuple` are compatible with
    these protocols
- Add runtime_checkable decorator to Protocols
  - This allows the Protocolcs to be checked by runtime type checkers
    such as beartype which can be used to verify jaxtyping annotations
