+++ dynamax/hidden_markov_model/models/abstractions.py
@@ -17,10 +17,11 @@ from jax.tree_util import tree_map
 from jaxtyping import Float, Array, PyTree
 import optax
 from tensorflow_probability.substrates.jax import distributions as tfd
-from typing import Any, Optional, Tuple
+from typing import Any, Optional, Tuple, runtime_checkable
 from typing_extensions import Protocol
 
 
+@runtime_checkable
 class HMMParameterSet(Protocol):
     """Container for HMM parameters.
 
@@ -28,11 +29,20 @@ class HMMParameterSet(Protocol):
     :param transitions: (ParameterSet) transition distribution parameters
     :param emissions: (ParameterSet) emission distribution parameters
     """
-    initial: ParameterSet
-    transitions: ParameterSet
-    emissions: ParameterSet
+    @property
+    def initial(self) -> ParameterSet:
+        pass
+
+    @property
+    def transitions(self) -> ParameterSet:
+        pass
+
+    @property
+    def emissions(self) -> ParameterSet:
+        pass
 
 
+@runtime_checkable
 class HMMPropertySet(Protocol):
     """Container for properties of HMM parameter properties.
 
@@ -40,10 +50,17 @@ class HMMPropertySet(Protocol):
     :param transitions: (PropertySet) transition distribution properties
     :param emissions: (PropertySet) emission distribution properties
     """
-    initial: PropertySet
-    transitions: PropertySet
-    emissions: PropertySet
+    @property
+    def initial(self) -> PropertySet:
+        pass
 
+    @property
+    def transitions(self) -> PropertySet:
+        pass
+
+    @property
+    def emissions(self) -> PropertySet:
+        pass
 
 
 class HMMInitialState(ABC):
