+++ dynamax/hidden_markov_model/models/logreg_hmm.py
@@ -138,7 +138,7 @@ class LogisticRegressionHMM(HMM):
         return (self.inputs_dim,)
 
     def initialize(self,
-                   key: jr.PRNGKey=jr.PRNGKey(0),
+                   key: Array=jr.PRNGKey(0),
                    method: str="prior",
                    initial_probs: Optional[Float[Array, " num_states"]]=None,
                    transition_matrix: Optional[Float[Array, "num_states num_states"]]=None,
