+++ dynamax/hidden_markov_model/models/linreg_hmm.py
@@ -184,7 +184,7 @@ class LinearRegressionHMM(HMM):
         return (self.input_dim,)
 
     def initialize(self,
-                   key: jr.PRNGKey=jr.PRNGKey(0),
+                   key: Array=jr.PRNGKey(0),
                    method: str="prior",
                    initial_probs: Optional[Float[Array, " num_states"]]=None,
                    transition_matrix: Optional[Float[Array, "num_states num_states"]]=None,
