### 2024-09-12 cfd3bfc
Fix type annotations in hmm parallel inference
