### 2024-09-13 dc7f675
Convert strings with latex to raw-strings
This stops type-checkers worrying about escape-sequences.
