### 2024-09-17 7da57ef
Minor arg and type changes in utils/utils.py
- Change default arguments to floats in `monotonically_increasing` function
- Help type-checkers parse `compute_state_overlap` function
- Update `keepdims` default value in `pytree_sum` to match `jnp.sum`
