+++ dynamax/utils/utils.py
@@ -43,7 +43,7 @@ def pad_sequences(observations, valid_lens, pad_val=0):
     return dataset
 
 
-def monotonically_increasing(x, atol=0, rtol=0):
+def monotonically_increasing(x, atol=0., rtol=0.):
     thresh = atol + rtol*jnp.abs(x[:-1])
     return jnp.all(jnp.diff(x) >= -thresh)
 
@@ -55,7 +55,7 @@ def pytree_len(pytree):
         return len(tree_leaves(pytree)[0])
 
 
-def pytree_sum(pytree, axis=None, keepdims=None, where=None):
+def pytree_sum(pytree, axis=None, keepdims=False, where=None):
     return tree_map(partial(jnp.sum, axis=axis, keepdims=keepdims, where=where), pytree)
 
 
@@ -166,7 +166,7 @@ def compute_state_overlap(
     assert z1.shape == z2.shape
     assert z1.min() >= 0 and z2.min() >= 0
 
-    K = max(z1.max(), z2.max()) + 1
+    K = max(max(z1), max(z2)) + 1
 
     overlap = jnp.sum(
         (z1[:, None] == jnp.arange(K))[:, :, None]
