### 2024-09-18 4c84ddc
Add further type annotations to categorical hmm
