### 2024-09-20 7e04c9e
Add further type annotations to arhmm
