### 2024-09-20 b2fffa4
Add further type annotations to linreghmm
