+++ dynamax/hidden_markov_model/models/linreg_hmm.py
@@ -1,7 +1,11 @@
+from typing import Any, Dict, NamedTuple, Optional, Tuple, Union
 import jax.numpy as jnp
 import jax.random as jr
 from jax import vmap
-from jaxtyping import Float, Array
+from jaxtyping import Array, Float, Int
+from tensorflow_probability.substrates import jax as tfp
+
+from dynamax.hidden_markov_model.inference import HMMPosterior
 from dynamax.hidden_markov_model.models.abstractions import HMM, HMMEmissions, HMMParameterSet, HMMPropertySet
 from dynamax.hidden_markov_model.models.initial import StandardHMMInitialState, ParamsStandardHMMInitialState
 from dynamax.hidden_markov_model.models.transitions import StandardHMMTransitions, ParamsStandardHMMTransitions
@@ -9,8 +13,6 @@ from dynamax.parameters import ParameterProperties
 from dynamax.types import Scalar
 from dynamax.utils.utils import pytree_sum
 from dynamax.utils.bijectors import RealToPSDBijector
-from tensorflow_probability.substrates import jax as tfp
-from typing import NamedTuple, Optional, Tuple, Union
 
 tfd = tfp.distributions
 tfb = tfp.bijectors
@@ -29,9 +31,9 @@ class ParamsLinearRegressionHMM(NamedTuple):
 
 class LinearRegressionHMMEmissions(HMMEmissions):
     def __init__(self,
-                 num_states,
-                 input_dim,
-                 emission_dim):
+                 num_states: int,
+                 input_dim: int,
+                 emission_dim: int):
         """_summary_
 
         Args:
@@ -46,16 +48,17 @@ class LinearRegressionHMMEmissions(HMMEmissions):
         self.emission_dim = emission_dim
 
     @property
-    def emission_shape(self):
+    def emission_shape(self) -> Tuple[int]:
         return (self.emission_dim,)
 
     def initialize(self,
-                   key=jr.PRNGKey(0),
-                   method="prior",
-                   emission_weights=None,
-                   emission_biases=None,
-                   emission_covariances=None,
-                   emissions=None):
+                   key: Array=jr.PRNGKey(0),
+                   method: str="prior",
+                   emission_weights: Optional[Float[Array, "num_states emission_dim input_dim"]]=None,
+                   emission_biases: Optional[Float[Array, "num_states emission_dim"]]=None,
+                   emission_covariances: Optional[Float[Array, "num_states emission_dim emission_dim"]]=None,
+                   emissions: Optional[Float[Array, "num_timesteps emission_dim"]]=None
+                   ) -> Tuple[ParamsLinearRegressionHMMEmissions, ParamsLinearRegressionHMMEmissions]:
         if method.lower() == "kmeans":
             assert emissions is not None, "Need emissions to initialize the model with K-Means!"
             from sklearn.cluster import KMeans
@@ -87,16 +90,27 @@ class LinearRegressionHMMEmissions(HMMEmissions):
             covs=ParameterProperties(constrainer=RealToPSDBijector()))
         return params, props
 
-    def distribution(self, params, state, inputs):
+    def distribution(
+            self,
+            params: ParamsLinearRegressionHMMEmissions,
+            state: Union[int, Int[Array, ""]],
+            inputs: Float[Array, " input_dim"]
+    ):
         prediction = params.weights[state] @ inputs
         prediction +=  params.biases[state]
         return tfd.MultivariateNormalFullCovariance(prediction, params.covs[state])
 
-    def log_prior(self, params):
[…]
