### 2024-09-20 a45e61b
Add further type annotations to Bernoulli HMM
