### 2024-09-20 32db62a
Add further type annotations to Gamma HMM
