+++ dynamax/hidden_markov_model/models/gamma_hmm.py
@@ -18,27 +18,33 @@ class ParamsGammaHMMEmissions(NamedTuple):
 
 
 class GammaHMMEmissions(HMMEmissions):
-    def __init__(self,
-                 num_states,
-                 m_step_optimizer=optax.adam(1e-2),
-                 m_step_num_iters=50):
+    def __init__(
+        self,
+        num_states: int,
+        m_step_optimizer: optax.GradientTransformation = optax.adam(1e-2),
+        m_step_num_iters: int = 50,
+    ):
         super().__init__(m_step_optimizer=m_step_optimizer, m_step_num_iters=m_step_num_iters)
         self.num_states = num_states
 
     @property
-    def emission_shape(self):
+    def emission_shape(self) -> Tuple:
         return ()
 
-    def initialize(self,
-                   key=jr.PRNGKey(0),
-                   method="prior",
-                   emission_concentrations=None,
-                   emission_rates=None,
-                   emissions=None):
+    def initialize(
+        self,
+        key: Array = jr.PRNGKey(0),
+        method="prior",
+        emission_concentrations: Optional[Float[Array, " num_states"]] = None,
+        emission_rates: Optional[Float[Array, " num_states"]] = None,
+        emissions: Optional[Float[Array, " num_timesteps"]] = None,
+        # ) -> Tuple[ParamsGammaHMMEmissions, ParamsGammaHMMEmissions]:
+    ) -> Tuple[ParamsGammaHMMEmissions, ParamsGammaHMMEmissions]:
 
         if method.lower() == "kmeans":
             assert emissions is not None, "Need emissions to initialize the model with K-Means!"
             from sklearn.cluster import KMeans
+
             key, subkey = jr.split(key)  # Create a random seed for SKLearn.
             sklearn_key = jr.randint(subkey, shape=(), minval=0, maxval=2147483647)  # Max int32 value.
             km = KMeans(self.num_states, random_state=int(sklearn_key)).fit(emissions.reshape(-1, 1))
@@ -57,18 +63,19 @@ class GammaHMMEmissions(HMMEmissions):
         default = lambda x, x0: x if x is not None else x0
         params = ParamsGammaHMMEmissions(
             concentration=default(emission_concentrations, _emission_concentrations),
-            rate=default(emission_rates, _emission_rates))
+            rate=default(emission_rates, _emission_rates),
+        )
         props = ParamsGammaHMMEmissions(
             concentration=ParameterProperties(constrainer=tfb.Softplus()),
-            rate=ParameterProperties(constrainer=tfb.Softplus()))
+            rate=ParameterProperties(constrainer=tfb.Softplus()),
+        )
         return params, props
 
-    def log_prior(self, params):
+    def log_prior(self, params) -> float:
         return 0.0
 
-    def distribution(self, params, state, inputs=None):
-        return tfd.Gamma(concentration=params.concentration[state],
-                         rate=params.rate[state])
+    def distribution(self, params: ParamsGammaHMMEmissions, state, inputs=None) -> tfd.Distribution:
+        return tfd.Gamma(concentration=params.concentration[state], rate=params.rate[state])
 
 
 class ParamsGammaHMM(NamedTuple):
@@ -96,27 +103,35 @@ class GammaHMM(HMM):
     :param m_step_num_iters: number of optimizer steps per M-step.
 
     """
-    def __init__(self,
-                 num_states: int,
-                 initial_probs_concentration: Union[Scalar, Float[Array, " num_states"]]=1.1,
-                 transition_matrix_concentration: Union[Scalar, Float[Array, " num_states"]]=1.1,
-                 transition_matrix_stickiness: Scalar=0.0,
[…]
