### 2024-09-20 81138f8
Add further type annotations to Gaussian HMMs
- Add annotations to the *GaussianHMMEmssions classes
- Apply some small formatting changes to reduce line length in places
