+++ dynamax/hidden_markov_model/models/gaussian_hmm.py
@@ -1,15 +1,18 @@
+from typing import Any, Dict, NamedTuple, Optional, Tuple, Union
+from jax import vmap
 import jax.numpy as jnp
 import jax.random as jr
 import tensorflow_probability.substrates.jax.bijectors as tfb
 import tensorflow_probability.substrates.jax.distributions as tfd
-from jax import vmap
 from jaxtyping import Float, Array
 import optax
+
 from dynamax.parameters import ParameterProperties
+from dynamax.hidden_markov_model.inference import HMMPosterior
 from dynamax.hidden_markov_model.models.abstractions import HMM, HMMEmissions, HMMParameterSet, HMMPropertySet
 from dynamax.hidden_markov_model.models.initial import StandardHMMInitialState, ParamsStandardHMMInitialState
 from dynamax.hidden_markov_model.models.transitions import StandardHMMTransitions, ParamsStandardHMMTransitions
-from dynamax.types import Scalar
+from dynamax.types import IntScalar, Scalar
 from dynamax.utils.distributions import InverseWishart
 from dynamax.utils.distributions import NormalInverseGamma
 from dynamax.utils.distributions import NormalInverseWishart
@@ -17,7 +20,6 @@ from dynamax.utils.distributions import nig_posterior_update
 from dynamax.utils.distributions import niw_posterior_update
 from dynamax.utils.bijectors import RealToPSDBijector
 from dynamax.utils.utils import pytree_sum
-from typing import NamedTuple, Optional, Tuple, Union
 
 
 class ParamsGaussianHMMEmissions(NamedTuple):
@@ -28,19 +30,21 @@ class ParamsGaussianHMMEmissions(NamedTuple):
 class GaussianHMMEmissions(HMMEmissions):
 
     def __init__(self,
-                 num_states,
-                 emission_dim,
-                 emission_prior_mean=0.0,
-                 emission_prior_concentration=1e-4,
-                 emission_prior_scale=1e-4,
-                 emission_prior_extra_df=0.1):
-        """_summary_
+                 num_states: int,
+                 emission_dim: int,
+                 emission_prior_mean: Union[Scalar, Float[Array, " emission_dim"]] = 0.0,
+                 emission_prior_concentration: Scalar = 1e-4,
+                 emission_prior_scale: Union[Scalar, Float[Array, "emission_dim emission_dim"]] = 1e-4,
+                 emission_prior_extra_df: Scalar = 0.1):
+        """Initialize GaussianHMMEmissions.
 
         Args:
-            initial_probabilities (_type_): _description_
-            transition_matrix (_type_): _description_
-            emission_means (_type_): _description_
-            emission_covariance_matrices (_type_): _description_
+            num_states: number of discrete states
+            emission_dim: dimension of the emission vector
+            emission_prior_mean: prior mean for emissions
+            emission_prior_concentration: concentration parameter for the prior
+            emission_prior_scale: scale matrix for the prior
+            emission_prior_extra_df: extra degrees of freedom for the prior
         """
         self.num_states = num_states
         self.emission_dim = emission_dim
@@ -51,23 +55,42 @@ class GaussianHMMEmissions(HMMEmissions):
         self.emission_prior_df = emission_dim + emission_prior_extra_df
 
     @property
-    def emission_shape(self):
+    def emission_shape(self) -> Tuple[int]:
         return (self.emission_dim,)
 
-    def distribution(self, params, state, inputs=None):
+    def distribution(
+            self, 
+            params: ParamsGaussianHMMEmissions,
+            state: IntScalar,
+            inputs: Optional[Array] = None
+            ) -> tfd.Distribution:
         return tfd.MultivariateNormalFullCovariance(
             params.means[state], params.covs[state])
 
[…]
