### 2024-09-20 4e36520
Add further type annotations to gmhmms
- Add annotations to the *GaussianMixtureHMMEmssions classes
