+++ dynamax/hidden_markov_model/models/gmm_hmm.py
@@ -1,3 +1,4 @@
+from typing import Any, Dict, NamedTuple, Optional, Tuple, Union
 import jax.numpy as jnp
 import jax.random as jr
 import tensorflow_probability.substrates.jax.bijectors as tfb
@@ -10,16 +11,15 @@ from dynamax.utils.distributions import NormalInverseGamma
 from dynamax.utils.distributions import NormalInverseWishart
 from dynamax.utils.distributions import nig_posterior_update
 from dynamax.utils.distributions import niw_posterior_update
+from dynamax.hidden_markov_model.inference import HMMPosterior
 from dynamax.hidden_markov_model.models.abstractions import HMM, HMMEmissions, HMMParameterSet, HMMPropertySet
 from dynamax.hidden_markov_model.models.initial import StandardHMMInitialState, ParamsStandardHMMInitialState
 from dynamax.hidden_markov_model.models.transitions import StandardHMMTransitions, ParamsStandardHMMTransitions
 from dynamax.utils.bijectors import RealToPSDBijector
 from dynamax.utils.utils import pytree_sum
-from dynamax.types import Scalar
-from typing import NamedTuple, Optional, Tuple, Union
+from dynamax.types import IntScalar, Scalar
 
 
-# Types
 class ParamsGaussianMixtureHMMEmissions(NamedTuple):
     weights: Union[Float[Array, "state_dim num_components"], ParameterProperties]
     means: Union[Float[Array, "state_dim num_components emission_dim"], ParameterProperties]
@@ -48,14 +48,14 @@ class ParamsDiagonalGaussianMixtureHMM(NamedTuple):
 class GaussianMixtureHMMEmissions(HMMEmissions):
 
     def __init__(self,
-                 num_states,
-                 num_components,
-                 emission_dim,
-                 emission_weights_concentration=1.1,
-                 emission_prior_mean=0.,
-                 emission_prior_mean_concentration=1e-4,
-                 emission_prior_extra_df=1e-4,
-                 emission_prior_scale=0.1):
+                 num_states: int,
+                 num_components: int,
+                 emission_dim: int,
+                 emission_weights_concentration: Union[Scalar, Float[Array, " num_components"]]=1.1,
+                 emission_prior_mean: Union[Scalar, Float[Array, " emission_dim"]]=0.,
+                 emission_prior_mean_concentration: Scalar=1e-4,
+                 emission_prior_extra_df: Scalar=1e-4,
+                 emission_prior_scale: Union[Scalar, Float[Array, "emission_dim emission_dim"]]=0.1):
         self.num_states = num_states
         self.num_components = num_components
         self.emission_dim = emission_dim
@@ -69,12 +69,14 @@ class GaussianMixtureHMMEmissions(HMMEmissions):
     def emission_shape(self):
         return (self.emission_dim,)
 
-    def initialize(self, key=jr.PRNGKey(0),
-                   method="prior",
-                   emission_weights=None,
-                   emission_means=None,
-                   emission_covariances=None,
-                   emissions=None):
+    def initialize(self,
+                   key: Array=jr.PRNGKey(0),
+                   method: str="prior",
+                   emission_weights: Optional[Float[Array, "num_states num_components"]]=None,
+                   emission_means: Optional[Float[Array, "num_states num_components emission_dim"]]=None,
+                   emission_covariances: Optional[Float[Array, "num_states num_components emission_dim emission_dim"]]=None,
+                   emissions: Optional[Float[Array, "num_timesteps emission_dim"]]=None
+        ) -> Tuple[ParamsGaussianMixtureHMMEmissions, ParamsGaussianMixtureHMMEmissions]:
         if method.lower() == "kmeans":
[…]
