### 2024-09-21 4b7ee53
Add further type annotations to logreg hmm
