### 2024-09-21 e246606
Add further type annotations to multinomialhmm
