### 2024-09-23 e19748d
Add further type annotations to poisson hmm
