### 2024-09-23 42c5c22
Add IntScalar type
