+++ dynamax/types.py
@@ -1,6 +1,8 @@
 from typing import Union
-from jaxtyping import Array, Float
+from jaxtyping import Array, Float, Int
 
 PRNGKeyT = Array
 
 Scalar = Union[float, Float[Array, ""]] # python float or scalar jax device array with dtype float
+
+IntScalar = Union[int, Int[Array, ""]]
