### 2024-09-23 45134e8
Update type annotations in hmm base classes
