+++ dynamax/hidden_markov_model/models/abstractions.py
@@ -1,6 +1,8 @@
 from abc import abstractmethod, ABC
+from typing import Any, Optional, Tuple, runtime_checkable, Union 
+from typing_extensions import Protocol
 from dynamax.ssm import SSM
-from dynamax.types import Scalar
+from dynamax.types import IntScalar, Scalar
 from dynamax.parameters import to_unconstrained, from_unconstrained
 from dynamax.parameters import ParameterSet, PropertySet
 from dynamax.hidden_markov_model.inference import HMMPosterior
@@ -11,14 +13,12 @@ from dynamax.hidden_markov_model.inference import hmm_two_filter_smoother
 from dynamax.utils.optimize import run_gradient_descent
 from dynamax.utils.utils import pytree_slice
 import jax.numpy as jnp
-import jax.random as jr
 from jax import vmap
 from jax.tree_util import tree_map
-from jaxtyping import Float, Array, PyTree
+from jaxtyping import Float, Array, PyTree, Real
 import optax
 from tensorflow_probability.substrates.jax import distributions as tfd
-from typing import Any, Optional, Tuple, runtime_checkable
-from typing_extensions import Protocol
+
 
 
 @runtime_checkable
@@ -113,14 +113,14 @@ class HMMInitialState(ABC):
         """
         raise NotImplementedError
 
-    def _compute_initial_probs(self, params, inputs=None):
-        return self.initial_distribution(params, inputs).probs_parameter()
+    def _compute_initial_probs(self, params, inputs:Optional[Array] = None):
+        return self.distribution(params, inputs).probs_parameter()
 
     def collect_suff_stats(self,
                            params: ParameterSet,
                            posterior: HMMPosterior,
                            inputs: Optional[Float[Array, "num_timesteps input_dim"]]=None
-    ) -> PyTree:
+                           ) -> Tuple[Float[Array, " num_states"], Optional[Float[Array, " input_dim"]]]:
         """Collect sufficient statistics for updating the initial distribution parameters.
 
         Args:
@@ -152,7 +152,7 @@ class HMMInitialState(ABC):
                batch_stats: PyTree,
                m_step_state: Any,
                scale: float=1.0
-    ) -> ParameterSet:
+    ) -> Tuple[ParameterSet, Any]:
         """Perform an M-step on the initial distribution parameters.
 
         Args:
@@ -211,7 +211,7 @@ class HMMTransitions(ABC):
     @abstractmethod
     def distribution(self,
                      params: ParameterSet,
-                     state: int,
+                     state: IntScalar,
                      inputs: Optional[Float[Array, " input_dim"]]=None
     ) -> tfd.Distribution:
         """Return a distribution over the next latent state
@@ -255,7 +255,7 @@ class HMMTransitions(ABC):
         """
         raise NotImplementedError
 
-    def _compute_transition_matrices(self, params, inputs=None):
+    def _compute_transition_matrices(self, params, inputs:Optional[Array] = None):
         if inputs is not None:
             f = lambda inpt: \
                 vmap(lambda state: \
@@ -271,7 +271,7 @@ class HMMTransitions(ABC):
                            params: ParameterSet,
                            posterior: HMMPosterior,
                            inputs: Optional[Float[Array, "num_timesteps input_dim"]]=None
-    ) -> PyTree:
+                           ) -> Tuple[Float[Array, "..."], Optional[Float[Array, "num_timesteps-1 input_dim"]]]:
         """Collect sufficient statistics for updating the transition distribution parameters.
 
         Args:
@@ -301,7 +301,7 @@ class HMMTransitions(ABC):
                batch_stats: PyTree,
                m_step_state: Any,
                scale: float=1.0
-    ) -> ParameterSet:
[…]
