### 2024-09-25 ee3da88
Add further type annotations to hmm transitions class
