### 2024-09-25 45e3e7f
Add further type annotations to hmm initial base class
