### 2024-10-06 076350a
fixup! Add further type annotations to linreghmm
