### 2024-10-06 c2098cf
Add further type annotations to categorical glm hmm
