### 2024-10-15 e06b594
Add __repr__ method to ParameterProperties
