### 2024-10-16 4b2c40b
Add further type annotations to hmm initial base class
Further, enforce that either key or initial_probs specified in
initialize method.
