### 2024-10-28 probml/dynamax#384 @slinderman
Bug in HMC notebook 
There is a bug in the [LGSSM + HMC notebook](https://probml.github.io/dynamax/notebooks/linear_gaussian_ssm/lgssm_hmc.html). It looks like the BlackJax interface for window adaptation schemes has changed slightly, and we need to update our code accordingly.
