### 2024-10-28 #7069 @slinderman
Hi @osorensen, sorry for the delay. We pinned Numpy < 2.0 and recreated the documentation, and now almost all of the rendering issues should be fixed. There is a known issue with the HMC notebook (see https://github.com/probml/dynamax/issues/384), but the rest look good.
