### 2024-11-01 probml/dynamax#377 @gdalle
I tried again to run the code on the documentation home page and got the following error:

```python
import jax.numpy as jnp
import jax.random as jr
from dynamax.hidden_markov_model import GaussianHMM

key1, key2, key3 = jr.split(jr.PRNGKey(0), 3)
num_states = 3
emission_dim = 2
num_timesteps = 1000

# Make a Gaussian HMM and sample data from it
hmm = GaussianHMM(num_states, emission_dim)
true_params, _ = hmm.initialize(key1)
```

```python
Traceback (most recent call last):
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/test.py", line 12, in <module>
    true_params, _ = hmm.initialize(key1)
                     ^^^^^^^^^^^^^^^^^^^^
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/.venv/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/gaussian_hmm.py", line 649, in initialize
    params["initial"], props["initial"] = self.initial_component.initialize(key1, method=method, initial_probs=initial_probs)
                                          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/.venv/lib/python3.12/site-packages/dynamax/hidden_markov_model/models/initial.py", line 45, in initialize
    initial_probs = tfd.Dirichlet(self.initial_probs_concentration).sample(seed=this_key)
                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py", line 1205, in sample
    return self._call_sample_n(sample_shape, seed, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/distribution.py", line 1182, in _call_sample_n
    samples = self._sample_n(
              ^^^^^^^^^^^^^^^
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/dirichlet.py", line 233, in _sample_n
    log_gamma_sample = gamma_lib.random_gamma(
                       ^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/gamma.py", line 725, in random_gamma
    return random_gamma_with_runtime(
           ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/distributions/gamma.py", line 718, in random_gamma_with_runtime
    seed = samplers.sanitize_seed(seed, salt='random_gamma')
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/internal/samplers.py", line 144, in sanitize_seed
    seed = fold_in(seed, salt)
           ^^^^^^^^^^^^^^^^^^^
  File "/home/gdalle/Work/GitHub/Python/dynamax-test/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/internal/samplers.py", line 186, in fold_in
    seed, jnp.asarray(salt & np.uint32(2**32 - 1), dtype=SEED_DTYPE))
                      ~~~~~^~~~~~~~~~~~~~~~~~~~~~
OverflowError: Python int too large to convert to C long
```
