### 2024-11-01 #7069 @gdalle
@slinderman thanks for the fixes. Can you tag a new release containing them, so that running the documentation examples does not require cloning the repo?
