### 2024-11-01 probml/dynamax#377 @slinderman
Hi @gdalle, I just pushed 0.1.5 to pypi and tagged it as the latest release. It has the numpy < 2.0 requirement and the test fixes. Sorry for missing this earlier!
