### 2024-11-01 #7069 @slinderman
Hi @gdalle, of course. Please see https://github.com/probml/dynamax/releases/tag/0.1.5, which is now available on [PyPI](https://pypi.org/project/dynamax/) as well.
