### 2024-11-06 f61c71b
Add inputs to parallel Kalman filter and smoother
