+++ dynamax/linear_gaussian_ssm/parallel_inference.py
@@ -33,7 +33,7 @@ Y₀             Y₁             Y₂             Y₃
 import jax.numpy as jnp
 from jax import vmap, lax
 from jaxtyping import Array, Float
-from typing import NamedTuple
+from typing import NamedTuple, Optional
 from dynamax.types import PRNGKey
 from functools import partial
 import warnings
@@ -45,6 +45,7 @@ from tensorflow_probability.substrates.jax.distributions import (
 from jax.scipy.linalg import cho_solve, cho_factor
 from dynamax.utils.utils import symmetrize, psd_solve
 from dynamax.linear_gaussian_ssm import PosteriorGSSMFiltered, PosteriorGSSMSmoothed, ParamsLGSSM
+from dynamax.linear_gaussian_ssm.inference import _zeros_if_none
 
 
 def _get_one_param(x, dim, t):
@@ -56,14 +57,16 @@ def _get_one_param(x, dim, t):
     else:
         return x
 
-def _get_params(params, num_timesteps, t):
+def _get_params(params: ParamsLGSSM, num_timesteps, t):
     """Helper function to get parameters at time t."""
     assert not callable(params.emissions.cov), "Emission covariance cannot be a callable."
 
     F = _get_one_param(params.dynamics.weights, 2, t)
+    B = _get_one_param(params.dynamics.input_weights, 2, t)
     b = _get_one_param(params.dynamics.bias, 1, t)
     Q = _get_one_param(params.dynamics.cov, 2, t)
     H = _get_one_param(params.emissions.weights, 2, t+1)
+    D = _get_one_param(params.emissions.input_weights, 2, t+1)
     d = _get_one_param(params.emissions.bias, 1, t+1)
 
     if len(params.emissions.cov.shape) == 1: 
@@ -81,7 +84,7 @@ def _get_params(params, num_timesteps, t):
             "The covariance will be interpreted as static and non-diagonal. To "
             "specify a dynamic and diagonal covariance, pass it as a 3D array.")
 
-    return F, b, Q, H, d, R
+    return F, B, b, Q, H, D, d, R
 
 
 #---------------------------------------------------------------------------#
@@ -151,13 +154,18 @@ class FilterMessage(NamedTuple):
     logZ: Float[Array, "ntime"]
 
 
-def _initialize_filtering_messages(params, emissions):
+def _initialize_filtering_messages(
+    params: ParamsLGSSM,
+    emissions: Float[Array, "ntime emission_dim"],
+    inputs: Optional[Float[Array, "ntime input_dim"]]=None
+    ):
     """Preprocess observations to construct input for filtering assocative scan."""
 
     num_timesteps = emissions.shape[0]
+    inputs = _zeros_if_none(inputs, (num_timesteps, 0))
     
-    def _first_message(params, y):
-        H, d, R = _get_params(params, num_timesteps, -1)[3:]
+    def _first_message(params, y, u):
+        H, D, d, R = _get_params(params, num_timesteps, -1)[4:]
         m = params.initial.mean
         P = params.initial.cov
 
@@ -166,34 +174,34 @@ def _initialize_filtering_messages(params, emissions):
         K = P @ H.T @ S_inv
 
         A = jnp.zeros_like(P)
-        b = m + K @ (y - H @ m - d)
+        b = m + K @ (y - H @ m - D @ u - d)
         C = symmetrize(P - K @ S @ K.T)
         eta = jnp.zeros_like(b)
         J = jnp.eye(len(b))
 
-        logZ = _marginal_loglik_elem(P, H, R, y)
[…]
