### 2024-11-08 probml/dynamax#377 @gdalle
Ok I finally managed to run the example code without issue on 0.1.5!

Here are a few remarks (keeping in mind that I don't usually code in Python):

- What is the percentage of source code covered by your test suite? It should be easy to figure out with tools such as https://github.com/nedbat/coveragepy, and it could help you detect untested code.
