### 2024-11-08 probml/dynamax#376 @gdalle
Making a second pass on the documentation, now that the errors have been fixed (the remarks above still apply).

## General feedback

Plots need more descriptions and axis labeling, especially when none of the axes represents time or iterations.

## HMMs

**casino HMM: Learning**

> As you can see, stochastic gradient descent converges much more quickly that full-batch gradient descent in this example. Intuitively, that’s because SGD takes multiple steps per epoch (i.e. each complete sweep through the dataset), whereas full-batch gradient descent takes only one.
> The algorithms appear to have converged, but have they learned the correct parameters? Let’s see…
> ...
> Ok, but not perfect!

The parameters learned by gradient descent are actually... pretty bad?
The second row of the transition matrix is way off, and the loadedness of one of the dies goes basically undetected. Is it just a bad initialization?

## Linear Gaussian SSMs

**Tracking an object using the Kalman filter**

> Sample some data from the model

It would be good to indicate the flow of time in this plot, which is not the horizontal axis.

> Perform online filtering / Perform offline smoothing

What are the black circles?
These plots need more explanations and not just code.

**Online linear regression using Kalman filtering**

The final plot needs more explanations with words.

**Parallel filtering and smoothing in an LG-SSM**

I only just noticed that non-colorblind people probably see both serial and parallel curves superposed due to the different colors? But as a colorblind person I can't differentiate between red and green here. So maybe you could shift one of the curves slightly, to highlight that they follow each other?

**MAP parameter estimation for an LG-SSM using EM and SGD**

> Data

What does the plot represent?

**Bayesian parameter estimation for an LG-SSM using HMC**

> Call HMC

This example errors. Perhaps an insufficiently tight version bound on `blackjax`?

```
TypeError: kernel() got an unexpected keyword argument 'num_steps'
```

## Nonlinear Gaussian SSMs

**Tracking a spiraling object using the extended / unscented Kalman filter**

> We now show how to do this using a simple nonlinear Gaussian SSM, combined with various extensions of the Kalman filter algorithm.

What is the difference between these variants (e.g. unscented vs extended Kalman filter)? How does the user choose between them?

> $q_t \in R^2$ [...] $r_t \in N(0, R)$

It would be clearer with different fonts for the set of real numbers $\mathbb{R}$ and the covariance matrix $R$.

> Extended Kalman filter

Making one of the two lines dotted would help with readability. As would making the larger circles red instead of black.

> Unscented Kalman filter

What changed between the two plots? How come the curves are suddenly superposed? We're missing a few explanations here.

**Tracking a 1d pendulum using Extended / Unscented Kalman filter/ smoother**

> Sample data and plot it

Why does it look like the measurements are thresholded instead of just randomly distributed around the true angle?

**Online learning for an MLP using extended Kalman filtering**

> Plot results

Again, the plots need more explanations, especially when the code is so verbose. What are the axes? What are the transparent curves, and the solid blue one? What are the dots?

## Generalized Gaussian SSMs

_Skipped for now_
