### 2024-11-08 #7069 @gdalle
Thank you for fixing the versions, I am now able to install and run tests smoothly.

I have made a second pass on the documentation[^1] and now that they are actually working, your tutorials show a huge effort in terms of coding and visualization! As stated in the issue, I feel like the user experience could be further improved with two additions:

[^1]: I only skipped the last section (Generalized Gaussian SSMs) and the API reference.

- a general overview of the package and its components (in order to not have to dig into the full API reference when I'm wondering how to do something)
- verbose explanations in each tutorial (with words and not just code) + better axis labeling on the plots

Of course I also have more specific remarks which are listed in the issues below:

- https://github.com/probml/dynamax/issues/386
- https://github.com/probml/dynamax/issues/377
- https://github.com/probml/dynamax/issues/376

Do you think such changes are reasonable requests? I know that this review has been going on for a while but I really think this will make the package much more user-friendly.
