### 2024-11-12 #7069 @osorensen
Thanks a lot for your review, @gdalle.

> Do you think such changes are reasonable requests?

As editor, I suggest @slinderman adresses as much as possible of the issues raised here. If there are certain things which may be too much work or out of scope, please let us know in this thread, and we can discuss it.
