### 2025-01-27 9d68595
run tests on python 3.10 through 3.13, since those are the versions supported by JAX
