+++ .github/workflows/run_tests.yml
@@ -16,7 +16,7 @@ jobs:
       matrix:
         # Select the Python versions to test against
         os: ["ubuntu-latest", "macos-latest"]
-        python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.10", "3.11", "3.12", "3.13"]
     name: ${{ matrix.os }} with Python ${{ matrix.python-version }}
     steps:
     - name: Clone the reference repository
