### 2025-02-12 #7069 @slinderman
Hi @osorensen, we've made some progress have incorporated one major PR to implement the typing and code coverage suggestions from the reviewers. We still have some work to go, but I think I would benefit from a hard deadline. What if we set end of February as our date to get this done, and we'll do as much as we can in the next two weeks?
