### 2025-02-17 3269acc
adding docstrings to HMM modules
