+++ dynamax/hidden_markov_model/models/test_models.py
@@ -1,10 +1,12 @@
-import pytest
+"""Tests for the HMM models."""
+
+import dynamax.hidden_markov_model as models
 import jax.numpy as jnp
 import jax.random as jr
-from jax import vmap
-import dynamax.hidden_markov_model as models
-from dynamax.utils.utils import monotonically_increasing
+import pytest
 
+from jax import vmap
+from dynamax.utils.utils import monotonically_increasing
 
 NUM_TIMESTEPS = 50
 
@@ -29,6 +31,7 @@ CONFIGS = [
 
 @pytest.mark.parametrize(["cls", "kwargs", "inputs"], CONFIGS)
 def test_sample_and_fit(cls, kwargs, inputs):
+    """Test that we can sample from and fit a model."""
     hmm = cls(**kwargs)
     key1, key2 = jr.split(jr.PRNGKey(42))
     params, param_props = hmm.initialize(key1)
@@ -40,6 +43,7 @@ def test_sample_and_fit(cls, kwargs, inputs):
 
 ## A few model-specific tests
 def test_categorical_hmm_viterbi():
+    """Test that the Viterbi algorithm works for a simple CategoricalHMM."""
     # From http://en.wikipedia.org/wiki/Viterbi_algorithm:
     hmm = models.CategoricalHMM(num_states=2, emission_dim=1, num_classes=3)
     params, props = hmm.initialize(
@@ -54,6 +58,7 @@ def test_categorical_hmm_viterbi():
 
 
 def test_gmm_hmm_vs_gmm_diag_hmm(key=jr.PRNGKey(0), num_states=4, num_components=3, emission_dim=2):
+    """Test that a GaussianMixtureHMM and DiagonalGaussianMixtureHMM are equivalent."""
     key1, key2, key3 = jr.split(key, 3)
     diag_hmm = models.DiagonalGaussianMixtureHMM(num_states, num_components, emission_dim)
     diag_params, _ = diag_hmm.initialize(key1)
@@ -87,6 +92,7 @@ def test_gmm_hmm_vs_gmm_diag_hmm(key=jr.PRNGKey(0), num_states=4, num_components
 
 
 def test_sample_and_fit_arhmm():
+    """Test that we can sample from and fit a LinearAutoregressiveHMM."""
     arhmm = models.LinearAutoregressiveHMM(num_states=4, emission_dim=2, num_lags=1)
     #key1, key2 = jr.split(jr.PRNGKey(int(datetime.now().timestamp())))
     key1, key2 = jr.split(jr.PRNGKey(42))
