### 2025-02-17 772b1cb
starting to add commit messages to lgssm
