### 2025-02-17 dd76041
adding docstrings for lgssm
