+++ dynamax/linear_gaussian_ssm/models.py
@@ -1,11 +1,15 @@
+"""
+Linear Gaussian State Space Models (LG-SSM) with and without conjugate priors.
+"""
+import jax.numpy as jnp
+import jax.random as jr
+import tensorflow_probability.substrates.jax.distributions as tfd
+
 from fastprogress.fastprogress import progress_bar
 from functools import partial
 from jax import jit
-import jax.numpy as jnp
-import jax.random as jr
 from jax.tree_util import tree_map
 from jaxtyping import Array, Float
-import tensorflow_probability.substrates.jax.distributions as tfd
 from tensorflow_probability.substrates.jax.distributions import MultivariateNormalFullCovariance as MVN
 from typing import Any, Optional, Tuple, Union, runtime_checkable
 from typing_extensions import Protocol 
@@ -64,14 +68,11 @@ class LinearGaussianSSM(SSM):
     :param has_emissions_bias:  Whether model contains an offset term $d$. Defaults to True.
 
     """
-    def __init__(
-        self,
-        state_dim: int,
-        emission_dim: int,
-        input_dim: int=0,
-        has_dynamics_bias: bool=True,
-        has_emissions_bias: bool=True
-    ):
+    def __init__(self,
+                 state_dim: int,
+                 emission_dim: int,
+                 input_dim: int=0,
+                 has_dynamics_bias: bool=True,has_emissions_bias: bool=True):
         self.state_dim = state_dim
         self.emission_dim = emission_dim
         self.input_dim = input_dim
@@ -80,26 +81,27 @@ class LinearGaussianSSM(SSM):
 
     @property
     def emission_shape(self):
+        """Shape of the emission vector."""
         return (self.emission_dim,)
 
     @property
     def inputs_shape(self):
+        """Shape of the input vector."""
         return (self.input_dim,) if self.input_dim > 0 else None
 
-    def initialize(
-        self,
-        key: PRNGKeyT =jr.PRNGKey(0),
-        initial_mean: Optional[Float[Array, " state_dim"]]=None,
-        initial_covariance=None,
-        dynamics_weights=None,
-        dynamics_bias=None,
-        dynamics_input_weights=None,
-        dynamics_covariance=None,
-        emission_weights=None,
-        emission_bias=None,
-        emission_input_weights=None,
-        emission_covariance=None
-    ) -> Tuple[ParamsLGSSM, ParamsLGSSM]:
+    def initialize(self,
+                   key: PRNGKeyT =jr.PRNGKey(0),
+                   initial_mean: Optional[Float[Array, " state_dim"]]=None,
+                   initial_covariance=None,
+                   dynamics_weights=None,
+                   dynamics_bias=None,
+                   dynamics_input_weights=None,
+                   dynamics_covariance=None,
+                   emission_weights=None,
+                   emission_bias=None,
+                   emission_input_weights=None,
+                   emission_covariance=None) \
+                   -> Tuple[ParamsLGSSM, ParamsLGSSM]:
         r"""Initialize model parameters that are set to None, and their corresponding properties.
 
         Args:
@@ -118,7 +120,6 @@ class LinearGaussianSSM(SSM):
         Returns:
             Tuple[ParamsLGSSM, ParamsLGSSM]: parameters and their properties.
         """
-
         # Arbitrary default values, for demo purposes.
         _initial_mean = jnp.zeros(self.state_dim)
         _initial_covariance = jnp.eye(self.state_dim)
@@ -169,86 +170,134 @@ class LinearGaussianSSM(SSM):
             )
         return params, props
 
-    def initial_distribution(
-        self,
-        params: ParamsLGSSM,
-        inputs: Optional[Float[Array, "ntime input_dim"]]=None
-    ) -> tfd.Distribution:
+    def initial_distribution(self,
+                             params: ParamsLGSSM,
[…]
