### 2025-02-17 4954fb5
adding docs for nonlinear gaussian ssm modules
