### 2025-02-17 7324b7f
adding docstrings for GGSSM
