### 2025-02-17 5f41a24
adding docstrings for utils
