+++ dynamax/utils/distributions.py
@@ -1,3 +1,6 @@
+"""
+This module contains custom distributions for Bayesian inference.
+"""
 import jax.numpy as jnp
 from jax import vmap
 from jax.scipy.linalg import solve_triangular
@@ -10,35 +13,29 @@ from dynamax.utils.utils import psd_solve
 
 
 class InverseWishart(tfd.TransformedDistribution):
+    r"""Implementation of an inverse Wishart distribution as a transformation of
+    a Wishart distribution. This distribution is defined by a scalar degrees of
+    freedom `df` and a scale matrix, `scale`.
 
-    def __new__(cls, *args, **kwargs):
-        # Patch for tfp 0.18.0.
-        # See https://github.com/tensorflow/probability/issues/1617
-        return tfd.Distribution.__new__(cls)
+    #### Mathematical Details
+    The probability density function (pdf) is,
+    ```none
+    pdf(X; df, scale) = det(X)**(-0.5 (df+k+1)) exp(-0.5 tr[inv(X) scale]) / Z
+    Z = 2**(0.5 df k) Gamma_k(0.5 df) |det(scale)|**(-0.5 df)
+    ```
 
+    where
+    * `df >= k` denotes the degrees of freedom,
+    * `scale` is a symmetric, positive definite, `k x k` matrix,
+    * `Z` is the normalizing constant, and,
+    * `Gamma_k` is the [multivariate Gamma function](
+        https://en.wikipedia.org/wiki/Multivariate_gamma_function).
+
+    Args:
+        df (_type_): _description_
+        scale (_type_): _description_
+    """
     def __init__(self, df, scale):
-        r"""Implementation of an inverse Wishart distribution as a transformation of
-        a Wishart distribution. This distribution is defined by a scalar degrees of
-        freedom `df` and a scale matrix, `scale`.
-
-        #### Mathematical Details
-        The probability density function (pdf) is,
-        ```none
-        pdf(X; df, scale) = det(X)**(-0.5 (df+k+1)) exp(-0.5 tr[inv(X) scale]) / Z
-        Z = 2**(0.5 df k) Gamma_k(0.5 df) |det(scale)|**(-0.5 df)
-        ```
-
-        where
-        * `df >= k` denotes the degrees of freedom,
-        * `scale` is a symmetric, positive definite, `k x k` matrix,
-        * `Z` is the normalizing constant, and,
-        * `Gamma_k` is the [multivariate Gamma function](
-            https://en.wikipedia.org/wiki/Multivariate_gamma_function).
-
-        Args:
-            df (_type_): _description_
-            scale (_type_): _description_
-        """
         self._df = df
         self._scale = scale
         # Compute the Cholesky of the inverse scale to parameterize a
@@ -56,8 +53,17 @@ class InverseWishart(tfd.TransformedDistribution):
 
         self._parameters = dict(df=df, scale=scale)
 
+    def __new__(cls, *args, **kwargs):
+        """Patch for tfp 0.18.0.
+        See https://github.com/tensorflow/probability/issues/1617
+        """
+        return tfd.Distribution.__new__(cls)
+
     @classmethod
     def _parameter_properties(self, dtype, num_classes=None):
+        """
+        Returns a dictionary mapping parameter names to properties.
+        """
         return dict(
             # Annotations may optionally specify properties, such as `event_ndims`,
             # `default_constraining_bijector_fn`, `specifies_shape`, etc.; see
@@ -67,19 +73,23 @@ class InverseWishart(tfd.TransformedDistribution):
 
     @property
     def df(self):
+        """Return the degrees of freedom."""
         return self._df
 
     @property
     def scale(self):
+        """Return the scale matrix."""
         return self._scale
 
     def _mean(self):
[…]
