+++ dynamax/utils/utils.py
@@ -11,6 +11,7 @@ from scipy.optimize import linear_sum_assignment
 from jax.scipy.linalg import cho_factor, cho_solve
 
 def has_tpu():
+    """Check if the current device is a TPU."""
     try:
         return isinstance(jax.devices()[0], jaxlib.xla_extension.TpuDevice)
     except:
@@ -36,6 +37,7 @@ def pad_sequences(observations, valid_lens, pad_val=0):
     """
 
     def pad(seq, len):
+        """Pad a single sequence."""
         idx = jnp.arange(1, seq.shape[0] + 1)
         return jnp.where(idx <= len, seq, pad_val)
 
@@ -44,11 +46,13 @@ def pad_sequences(observations, valid_lens, pad_val=0):
 
 
 def monotonically_increasing(x, atol=0., rtol=0.):
+    """Check if an array is monotonically increasing."""
     thresh = atol + rtol*jnp.abs(x[:-1])
     return jnp.all(jnp.diff(x) >= -thresh)
 
 
 def pytree_len(pytree):
+    """Return the number of leaves in a PyTree."""
     if pytree is None:
         return 0
     else:
@@ -56,14 +60,17 @@ def pytree_len(pytree):
 
 
 def pytree_sum(pytree, axis=None, keepdims=False, where=None):
+    """Sum all the leaves in a PyTree."""
     return tree_map(partial(jnp.sum, axis=axis, keepdims=keepdims, where=where), pytree)
 
 
 def pytree_slice(pytree, slc):
+    """Slice all the leaves in a Pytree."""
     return tree_map(lambda x: x[slc], pytree)
 
 
 def pytree_stack(pytrees):
+    """Stack all the leaves in a list of PyTrees."""
     _, treedef = tree_flatten(pytrees[0])
     leaves = [tree_leaves(tree) for tree in pytrees]
     return tree_unflatten(treedef, [jnp.stack(vals) for vals in zip(*leaves)])
