### 2025-02-17 c43a3c9
adding docstrings for slds
