### 2025-02-17 d9b185e
cleaning up lgssm tests
