+++ dynamax/hidden_markov_model/inference_test.py
@@ -336,22 +336,18 @@ def test_parallel_posterior_sample(
         key = jr.PRNGKey(key)
 
     max_unique_size = 1 << num_timesteps
+    
+    k1, k2 = jr.split(key)
+    args = random_hmm_args(k1, num_timesteps, num_states)
 
-    def iterate_test(key_iter):
-        keys_iter = jr.split(key_iter, num_samples)
-        args = random_hmm_args(key_iter, num_timesteps, num_states)
+    # Sample sequences from posterior
+    state_seqs = vmap(parallel.hmm_posterior_sample, (0, None, None, None), (0, 0))(jr.split(k2, num_samples), *args)[1]
+    unique_seqs, counts = jnp.unique(state_seqs, axis=0, size=max_unique_size, return_counts=True)
+    blj_sample = counts / counts.sum()
 
-        # Sample sequences from posterior
-        state_seqs = vmap(parallel.hmm_posterior_sample, (0, None, None, None), (0, 0))(keys_iter, *args)[1]
-        unique_seqs, counts = jnp.unique(state_seqs, axis=0, size=max_unique_size, return_counts=True)
-        blj_sample = counts / counts.sum()
+    # Compute joint probabilities
+    blj = jnp.exp(big_log_joint(*args))
+    blj = jnp.ravel(blj / blj.sum())
 
-        # Compute joint probabilities
-        blj = jnp.exp(big_log_joint(*args))
-        blj = jnp.ravel(blj / blj.sum())
-
-        # Compare the joint distributions
-        return jnp.allclose(blj_sample, blj, rtol=0, atol=eps)
-
-    keys = jr.split(key, num_iterations)
-    assert iterate_test(keys[0])
+    # Compare the joint distributions
+    assert jnp.allclose(blj_sample, blj, rtol=0, atol=eps)
