### 2025-02-17 75cedea
fixing type bug. test cleanup exposed a bug in the parallel lgssm inference code
