### 2025-02-18 621027b
fixing bug in parallel lgssm inference with inputs
