+++ dynamax/linear_gaussian_ssm/parallel_inference.py
@@ -1,51 +1,46 @@
-'''
-Parallel filtering and smoothing for a lgssm.
+"""
+Parallel filtering and smoothing for a linear Gaussian SSM.
 
-This implementation is adapted from the work of Adrien Correnflos:
-https://github.com/EEA-sensors/sequential-parallelization-examples/
+The trick is to write the filtering and smoothing computations in 
+terms of a binary associative scan. This allows us to parallelize
+the computation of the filtering and smoothing messages.
 
-Note that in the original implementation, the initial state distribution
-applies to t=0, and the first emission occurs at time `t=1` (i.e. after
-the initial state has been transformed by the dynamics), whereas here,
-the first emission occurs at time `t=0` and is produced directly by the
-untransformed initial state (see below).
+The filtering messages represent two distributions:
+    
+    f_ij(x_i, x_j) = p(x_i | y_{j+1:i}, x_j)
+    g_ij(x_j) = p(y_i | x_j)
 
-Sarkka et al.
+We define a binary associative operator that takes two filtering messages
+and returns a new filtering message,
 
-      F₀,Q₀          F₁,Q₁         F₂,Q₂
-Z₀ ─────────── Z₁ ─────────── Z₂ ─────────── Z₃ ─────...
-               |              |              |
-               | H₁,R₁        | H₂,R₂        | H₃,R₃
-               |              |              |
-               Y₁             Y₂             Y₃
+    (f_ij, g_ij) \otimes (f_jk, g_jk) = (f_ik, g_ik)
 
-Dynamax
+For a linear Gaussian SSM, the filtering messages are Gaussian potentials.
+Following Sarkka and Garcia-Fernandez (2021), we write them in the following
+form,
+    
+    f_{ij}(x_i, x_j) = N(x_i | A_{ij} x_j + b_{ij}, C_{ij})
+    g_{ij}(x_j) \propto N_I(x_j | \eta_{ij}, J_{ij})
 
-      F₀,Q₀           F₁,Q₁         F₂,Q₂
-Z₀ ─────────── Z₁ ─────────── Z₂ ─────────── Z₃ ─────...
-|              |              |              |
-| H₀,R₀        | H₁,R₁        | H₂,R₂        | H₃,R₃
-|              |              |              |
-Y₀             Y₁             Y₂             Y₃ 
-
-'''
-
+where A_{ij}, b_{ij}, C_{ij}, \eta_{ij}, J_{ij} are the parameters of the
+Gaussian potentials. The updates for these parameters are derived in the 
+manuscript referenced above.
+"""
 import jax.numpy as jnp
-from jax import vmap, lax
-from jaxtyping import Array, Float
-from typing import NamedTuple, Optional
-from dynamax.types import PRNGKeyT
-from functools import partial
 import warnings
 
-from tensorflow_probability.substrates.jax.distributions import (
-    MultivariateNormalDiagPlusLowRankCovariance as MVNLowRank,
-    MultivariateNormalFullCovariance as MVN)
-
-from jax.scipy.linalg import cho_solve, cho_factor
 from dynamax.utils.utils import symmetrize, psd_solve
 from dynamax.linear_gaussian_ssm import PosteriorGSSMFiltered, PosteriorGSSMSmoothed, ParamsLGSSM
 from dynamax.linear_gaussian_ssm.inference import _zeros_if_none
+from dynamax.types import PRNGKeyT
+from functools import partial
+from jax import vmap, lax
+from jax.scipy.linalg import cho_solve, cho_factor
+from jaxtyping import Array, Float
+from tensorflow_probability.substrates.jax.distributions import (
[…]
