### 2025-02-18 a00339d
migrate to pyproject.toml
