### 2025-02-18 3f10061
updating versioneer to use pyproject.toml
