### 2025-02-18 49b3a5a
adding versioneer[toml] to pyproject.toml build-system requires
