### 2025-02-18 5aacae9
Update release.yml
Install versioneer[toml] before building the package.
