### 2025-02-18 c11eef8
Update pyproject.toml
Trying to fix the pypi release to include all the source files...
