### 2025-02-18 c7d6231
Update pyproject.toml
Changing versioneer style to hopefully be compatible with testpypi.
