### 2025-02-18 90053c6
Update release.yml
Only run this if explicitly triggered
