### 2025-02-18 8aff320
removing email from authors list
