### 2025-02-18 7cdc3a3
use python 3.11 for docs
