+++ .github/workflows/publish_documentation.yml
@@ -17,10 +17,10 @@ jobs:
         with:
           persist-credentials: false
 
-      - name: Set up Python 3.10
+      - name: Set up Python 3.11
         uses: actions/setup-python@v1
         with:
-          python-version: 3.10
+          python-version: 3.11
 
       - name: Build the documentation with Sphinx
         run: |
