### 2025-02-18 ad890ba
fixing some notebook bugs
