### 2025-02-18 b402c99
Adding typing for lgssm inference helper functions
