+++ dynamax/linear_gaussian_ssm/inference.py
@@ -158,7 +158,7 @@ def _get_one_param(x, dim, t):
         return x
 
 def _get_params(params, num_timesteps, t):
-    """Helper function to get parameters at time t."""
+    """Helper function to get all parameters at time t."""
     assert not callable(params.emissions.cov), "Emission covariance cannot be a callable."
 
     F = _get_one_param(params.dynamics.weights, 2, t)
@@ -199,8 +199,12 @@ def make_lgssm_params(initial_mean: Float[Array, " state_dim"],
                       dynamics_bias: Optional[Float[Array, " state_dim"]]=None,
                       dynamics_input_weights: Optional[Float[Array, "state_dim input_dim"]]=None,
                       emissions_bias: Optional[Float[Array, " emission_dim"]]=None,
-                      emissions_input_weights: Optional[Float[Array, "emission_dim input_dim"]]=None):
+                      emissions_input_weights: Optional[Float[Array, "emission_dim input_dim"]]=None
+                      ) -> ParamsLGSSM:
     """Helper function to construct a ParamsLGSSM object from arguments.
+
+    See `ParamsLGSSM`, `ParamsLGSSMInitial`, `ParamsLGSSMDynamics`, and `ParamsLGSSMEmissions` for
+    more details on the parameters.
     """
     state_dim = len(initial_mean)
     emission_dim = emissions_cov.shape[-1]
@@ -228,31 +232,37 @@ def make_lgssm_params(initial_mean: Float[Array, " state_dim"],
     return params
 
 
-def _predict(m, S, F, B, b, Q, u):
+def _predict(prior_mean: Float[Array, "state_dim"],
+             prior_cov: Float[Array, "state_dim state_dim"],
+             dynamics_matrix: Float[Array, "state_dim state_dim"],
+             input_weights: Float[Array, "state_dim input_dim"],
+             dynamics_bias: Float[Array, "state_dim"],
+             dynamics_cov: Float[Array, "state_dim state_dim"],
+             inpt: Float[Array, "input_dim"]
+             ) -> Tuple[Float[Array, "state_dim"], 
+                        Float[Array, "state_dim state_dim"]]:
     r"""Predict next mean and covariance under a linear Gaussian model.
 
         p(z_{t+1}) = int N(z_t \mid m, S) N(z_{t+1} \mid Fz_t + Bu + b, Q)
                     = N(z_{t+1} \mid Fm + Bu, F S F^T + Q)
 
-    Args:
-        m (D_hid,): prior mean.
-        S (D_hid,D_hid): prior covariance.
-        F (D_hid,D_hid): dynamics matrix.
-        B (D_hid,D_in): dynamics input matrix.
-        u (D_in,): inputs.
-        Q (D_hid,D_hid): dynamics covariance matrix.
-        b (D_hid,): dynamics bias.
-
     Returns:
-        mu_pred (D_hid,): predicted mean.
-        Sigma_pred (D_hid,D_hid): predicted covariance.
+        mu_pred (state_dim,): predicted mean.
+        Sigma_pred (state_dim,state_dim): predicted covariance.
     """
-    mu_pred = F @ m + B @ u + b
-    Sigma_pred = F @ S @ F.T + Q
+    mu_pred = dynamics_matrix @ prior_mean + input_weights @ inpt + dynamics_bias
+    Sigma_pred = dynamics_matrix @ prior_cov @ dynamics_matrix.T + dynamics_cov
     return mu_pred, Sigma_pred
 
 
-def _condition_on(m, P, H, D, d, R, u, y):
+def _condition_on(prior_mean: Float[Array, "state_dim"],
+                  prior_cov: Float[Array, "state_dim state_dim"],
[…]
