+++ dynamax/nonlinear_gaussian_ssm/inference_ukf.py
@@ -30,29 +30,29 @@ _process_input = lambda x, y: jnp.zeros((y,)) if x is None else x
 _compute_lambda = lambda x, y, z: x**2 * (y + z) - z
 
 
-def _compute_sigmas(m: Float[Array, "state_dim"], 
-                    P: Float[Array, "state_dim state_dim"], 
-                    n: int, 
-                    lamb: float) \
+def _compute_sigmas(mean: Float[Array, "state_dim"], 
+                    cov: Float[Array, "state_dim state_dim"], 
+                    dim: int, 
+                    scale: float) \
                     -> Float[Array, "2*state_dim+1"]:
     """Compute (2n+1) sigma points used for inputs to  unscented transform.
 
     Args:
-        m (D_hid,): mean.
-        P (D_hid,D_hid): covariance.
-        n (int): number of state dimensions.
+        mean (dim,): mean.
+        cov (dim, dim): covariance.
+        dim (int): number of state dimensions.
         lamb (Scalar): unscented parameter lambda.
 
     Returns:
-        sigmas (2*D_hid+1,): 2n+1 sigma points.
+        sigmas (2*dim+1,): 2dim+1 sigma points.
     """
-    distances = jnp.sqrt(n + lamb) * jnp.linalg.cholesky(P)
-    sigma_plus = jnp.array([m + distances[:, i] for i in range(n)])
-    sigma_minus = jnp.array([m - distances[:, i] for i in range(n)])
-    return jnp.concatenate((jnp.array([m]), sigma_plus, sigma_minus))
+    distances = jnp.sqrt(dim + scale) * jnp.linalg.cholesky(cov)
+    sigma_plus = jnp.array([mean + distances[:, i] for i in range(dim)])
+    sigma_minus = jnp.array([mean - distances[:, i] for i in range(dim)])
+    return jnp.concatenate((jnp.array([mean]), sigma_plus, sigma_minus))
 
 
-def _compute_weights(n: int, 
+def _compute_weights(dim: int, 
                      alpha: float, 
                      beta: float, 
                      lamb: float) \
@@ -61,7 +61,7 @@ def _compute_weights(n: int,
     """Compute weights used to compute predicted mean and covariance (Sarkka 5.77).
 
     Args:
-        n (int): number of state dimensions.
+        dim (int): number of state dimensions.
         alpha (float): hyperparameter that determines the spread of sigma points
         beta (float): hyperparameter that incorporates prior information
         lamb (float): lamb = alpha**2 *(n + kappa) - n
@@ -70,77 +70,71 @@ def _compute_weights(n: int,
         w_mean (2*n+1,): 2n+1 weights to compute predicted mean.
         w_cov (2*n+1,): 2n+1 weights to compute predicted covariance.
     """
-    factor = 1 / (2 * (n + lamb))
-    w_mean = jnp.concatenate((jnp.array([lamb / (n + lamb)]), jnp.ones(2 * n) * factor))
-    w_cov = jnp.concatenate((jnp.array([lamb / (n + lamb) + (1 - alpha**2 + beta)]), jnp.ones(2 * n) * factor))
+    factor = 1 / (2 * (dim + lamb))
+    w_mean = jnp.concatenate((jnp.array([lamb / (dim + lamb)]), jnp.ones(2 * dim) * factor))
+    w_cov = jnp.concatenate((jnp.array([lamb / (dim + lamb) + (1 - alpha**2 + beta)]), jnp.ones(2 * dim) * factor))
     return w_mean, w_cov
 
 
-def _predict(m: Float[Array, "state_dim"], 
-             P: Float[Array, "state_dim state_dim"],
-             f: Callable, 
[…]
