+++ dynamax/nonlinear_gaussian_ssm/inference_ekf.py
@@ -7,7 +7,7 @@ from jax import lax
 from jax import jacfwd
 from tensorflow_probability.substrates.jax.distributions import MultivariateNormalFullCovariance as MVN
 from jaxtyping import Array, Float
-from typing import List, Optional
+from typing import Callable, List, Optional, Tuple
 
 from dynamax.utils.utils import psd_solve, symmetrize
 from dynamax.nonlinear_gaussian_ssm.models import ParamsNLGSSM
@@ -20,31 +20,36 @@ _process_fn = lambda f, u: (lambda x, y: f(x)) if u is None else f
 _process_input = lambda x, y: jnp.zeros((y,1)) if x is None else x
 
 
-def _predict(m, P, f, F, Q, u):
+def _predict(prior_mean: Float[Array, " state_dim"], 
+             prior_cov: Float[Array, "state_dim state_dim"],
+             dynamics_func: Callable, 
+             dynamics_jacobian: Callable, 
+             dynamics_cov: Float[Array, "state_dim state_dim"],
+             inpt: Float[Array, " input_dim"]
+             ) -> Tuple[Float[Array, " state_dim"], Float[Array, "state_dim state_dim"]]:
     r"""Predict next mean and covariance using first-order additive EKF
 
         p(z_{t+1}) = \int N(z_t | m, S) N(z_{t+1} | f(z_t, u), Q)
                     = N(z_{t+1} | f(m, u), F(m, u) S F(m, u)^T + Q)
 
-    Args:
-        m (D_hid,): prior mean.
-        P (D_hid,D_hid): prior covariance.
-        f (Callable): dynamics function.
-        F (Callable): Jacobian of dynamics function.
-        Q (D_hid,D_hid): dynamics covariance matrix.
-        u (D_in,): inputs.
-
     Returns:
         mu_pred (D_hid,): predicted mean.
         Sigma_pred (D_hid,D_hid): predicted covariance.
     """
-    F_x = F(m, u)
-    mu_pred = f(m, u)
-    Sigma_pred = F_x @ P @ F_x.T + Q
+    F_x = dynamics_jacobian(prior_mean, inpt)
+    mu_pred = dynamics_func(prior_mean, inpt)
+    Sigma_pred = F_x @ prior_cov @ F_x.T + dynamics_cov
     return mu_pred, Sigma_pred
 
 
-def _condition_on(m, P, h, H, R, u, y, num_iter):
+def _condition_on(prior_mean: Float[Array, " state_dim"],
+                  prior_cov: Float[Array, "state_dim state_dim"],
+                  emission_func: Callable, 
+                  emission_jacobian: Callable, 
+                  emission_cov: Float[Array, "emission_dim emission_dim"],
+                  inpt: Float[Array, " input_dim"],
+                  emission: Float[Array, " emission_dim"],
+                  num_iter: int):
     r"""Condition a Gaussian potential on a new observation.
 
        p(z_t | y_t, u_t, y_{1:t-1}, u_{1:t-1})
@@ -59,16 +64,6 @@ def _condition_on(m, P, h, H, R, u, y, num_iter):
          SS = P - K * S * K' = Sigma_cond
      **Note! This can be done more efficiently when R is diagonal.**
 
-    Args:
-         m (D_hid,): prior mean.
-         P (D_hid,D_hid): prior covariance.
-         h (Callable): emission function.
-         H (Callable): Jacobian of emission function.
-         R (D_obs,D_obs): emission covariance matrix.
-         u (D_in,): inputs.
-         y (D_obs,): observation.
-         num_iter (int): number of re-linearizations around posterior for update step.
-
      Returns:
[…]
