### 2025-02-19 probml/dynamax#384 @slinderman
Fixed in https://github.com/probml/dynamax/commit/ad890ba0a0969994728375cf68b454fc8b5c6db4!
