+++ dynamax/linear_gaussian_ssm/models.py
@@ -7,12 +7,12 @@ import tensorflow_probability.substrates.jax.distributions as tfd
 
 from fastprogress.fastprogress import progress_bar
 from functools import partial
-from jax import jit, vmap
+from jax import jit, tree, vmap
 from jax.tree_util import tree_map
 from jaxtyping import Array, Float
 from tensorflow_probability.substrates.jax.distributions import MultivariateNormalFullCovariance as MVN
 from typing import Any, Optional, Tuple, Union, runtime_checkable
-from typing_extensions import Protocol 
+from typing_extensions import Protocol
 
 from dynamax.ssm import SSM
 from dynamax.linear_gaussian_ssm.inference import lgssm_joint_sample, lgssm_filter, lgssm_smoother, lgssm_posterior_sample
@@ -24,7 +24,7 @@ from dynamax.utils.bijectors import RealToPSDBijector
 from dynamax.utils.distributions import MatrixNormalInverseWishart as MNIW
 from dynamax.utils.distributions import NormalInverseWishart as NIW
 from dynamax.utils.distributions import mniw_posterior_update, niw_posterior_update
-from dynamax.utils.utils import pytree_stack, psd_solve
+from dynamax.utils.utils import ensure_array_has_batch_dim, pytree_stack, psd_solve
 
 @runtime_checkable
 class SuffStatsLGSSM(Protocol):
@@ -206,7 +206,7 @@ class LinearGaussianSSM(SSM):
                key: PRNGKeyT,
                num_timesteps: int,
                inputs: Optional[Float[Array, "num_timesteps input_dim"]] = None) \
-                -> Tuple[Float[Array, "num_timesteps state_dim"], 
+                -> Tuple[Float[Array, "num_timesteps state_dim"],
                          Float[Array, "num_timesteps emission_dim"]]:
         """Sample from the model.
         
@@ -357,7 +357,7 @@ class LinearGaussianSSM(SSM):
                 input_weights=params.emissions.input_weights,
                 cov=1e8 * jnp.ones(self.emission_dim)) # ignore dummy observatiosn
             )
-        
+
         dummy_emissions = jnp.zeros((num_forecast_timesteps, self.emission_dim))
         forecast_inputs = forecast_inputs if forecast_inputs is not None else \
             jnp.zeros((num_forecast_timesteps, 0))
@@ -367,7 +367,7 @@ class LinearGaussianSSM(SSM):
         H = params.emissions.weights
         b = params.emissions.bias
         R = params.emissions.cov if params.emissions.cov.ndim == 2 else jnp.diag(params.emissions.cov)
-        
+
         forecast_emissions = forecast_states.filtered_means @ H.T + b
         forecast_emissions_cov = H @ forecast_states.filtered_covariances @ H.T + R
         return forecast_states.filtered_means, \
@@ -662,18 +662,20 @@ class LinearGaussianConjugateSSM(LinearGaussianSSM):
         Returns:
             parameter object, where each field has `sample_size` copies as leading batch dimension.
         """
-        num_timesteps = len(emissions)
+        batch_emissions = ensure_array_has_batch_dim(emissions, self.emission_shape)
+        batch_inputs = ensure_array_has_batch_dim(inputs, self.inputs_shape)
 
-        if inputs is None:
-            inputs = jnp.zeros((num_timesteps, 0))
+        num_batches, num_timesteps = batch_emissions.shape[:2]
 
-        def sufficient_stats_from_sample(states):
+        if batch_inputs is None:
+            batch_inputs = jnp.zeros((num_batches, num_timesteps, 0))
+
+        def sufficient_stats_from_sample(y, inputs, states):
             """Convert samples of states to sufficient statistics."""
[…]
