### 2025-03-01 5202215
Add support for inhomogeneous parameters
