### 2025-03-01 e4f556a
Add support for inhomogeneous parameters
