### 2025-03-07 6ceb285
allowing for a bit of tolerance in the parallel log likelihood comparisons
