+++ dynamax/linear_gaussian_ssm/parallel_inference_test.py
@@ -145,8 +145,11 @@ class TestParallelLGSSMSmoother:
 
     def test_marginal_loglik(self):
         """Check if marginal log likelihoods are close."""
-        assert jnp.allclose(self.serial_posterior.marginal_loglik, self.parallel_posterior.marginal_loglik, atol=2e-1)
-        assert jnp.allclose(self.serial_posterior.marginal_loglik, self.parallel_posterior_diag.marginal_loglik, atol=2e-1)
+        ser_mll = self.serial_posterior.marginal_loglik
+        par_mll = self.parallel_posterior.marginal_loglik
+        par_diag_mll = self.parallel_posterior_diag.marginal_loglik
+        assert jnp.allclose(ser_mll, par_mll, atol=1e-4 * self.num_timesteps)
+        assert jnp.allclose(ser_mll, par_diag_mll, atol=1e-4 * self.num_timesteps)
 
 
 class TestParallelLGSSMSmootherWithInputs:
