### 2025-03-07 bfa83cb
allowing for a bit of tolerance in the parallel log likelihood comparisons
