### 2025-03-07 #7069 @slinderman
Hi @osorensen, @thomaspinder, and @gdalle,

Thank you for these incredibly helpful comments! I appreciate your patience while we've worked to incorporate your suggestions and reply.

---

@gdalle:
I just replied to your comments (in issues https://github.com/probml/dynamax/issues/376 and https://github.com/probml/dynamax/issues/377) on the documentation and code. I think the documentation in particular is much better now!

---

@thomaspinder:

Thanks for all of these suggestions. We have implemented the following:

1. Failing tests
- These are working on all of our automated tests (which now run on linux and Mac for our supported Python versions)

2. Docstring coverage
- We went through and added more extensive docstrings, especially for public-facing functions.
- We also added an automatic workflow to compute the docstring coverage, and we are now at 100% coverage!
- We also compute test coverage, and that is currently at 76% (though the number is dragged down by some old demo code that will eventually move to notebooks or be deleted).

3. Broken documentation
- We fixed the bugs in the documentation (major one was from numpy 2.0 dependency, which is now pinned).
- We set the doc workflow to fail upon notebook execution error

4. Incomplete contribution guidelines
- We added the info about how to commit and push

5. Missing typing
- We had typing for all of the functions documented in the API, and we added it to several of the other internal functions. (We are using JaxTyping.)
- We changed names of arguments to make them more interpretable, even for private functons like `_predict`
- We added the extra space for the single-dimension array typing specs

6.  Large repo size
- Thank you for this suggestion! `git gc` cleaned up the unnecessary files and now we're down to 7.3MB

7. Tighter Python bounds
- Great point. JAX itself has pretty narrow bounds (>= 3.10) so we set Dynamax to inherit that constraint.

8. Package management
- We switched to pyproject.toml for packaging. It was a bit of work to get it right, but I think this is big improvement!
