### 2025-03-14 fa1869e
updating paper per Guillaume's comments
