+++ paper/paper.md
@@ -62,14 +62,14 @@ There are a few key design choices to make when constructing an SSM:
 - How do the latent states evolve over time? E.g., are the dynamics linear or nonlinear?
 - How are the observations distributed? E.g., are they Gaussian, Poisson, etc.?
 
-Some design choices are so common they have their own names. Hidden Markov models (HMM) are SSMs with discrete latent states, and linear dynamical systems (LDS) are SSMs with continuous latent states, linear dynamics, and additive Gaussian noise.  `Dynamax` supports canonical SSMs and allows the user to construct bespoke models as needed, simply by inheriting from a base class and specifying a few model-specific functions.
+Some design choices are so common they have their own names. Hidden Markov models (HMM) are SSMs with discrete latent states, and linear dynamical systems (LDS) are SSMs with continuous latent states, linear dynamics, and additive Gaussian noise.  `Dynamax` supports canonical SSMs and allows the user to construct bespoke models as needed, simply by inheriting from a base class and specifying a few model-specific functions. For example, see the [_Creating Custom HMMs_](https://probml.github.io/dynamax/notebooks/hmm/custom_hmm.html) tutorial in the Dynamax documentation. 
 
 Finally, even for canonical models, there are several algorithms for state inference and parameter estimation. `Dynamax` provides robust implementations of several low-level inference algorithms to suit a variety of applications, allowing users to choose among a host of models and algorithms for their application. More information about state space models and algorithms for state inference and parameter estimation can be found in the textbooks by @murphy2023probabilistic and @sarkka2023bayesian. 
 
 
 # Statement of need
 
-`Dynamax` is an open-source Python package for state space modeling. Since it is built with `JAX` [@jax], it supports just-in-time (JIT) compilation for hardware acceleration on CPU, GPU, and TPU machines. It also supports automatic differentiation for gradient-based model learning. While other libraries exist for state space modeling in Python [@pyhsmm; @ssm; @eeasensors; @seabold2010statsmodels; @hmmlearn],  `Dynamax` provides a combination of low-level inference algorithms and high-level modeling objects that can support a wide range of research applications in JAX.
+`Dynamax` is an open-source Python package for state space modeling. Since it is built with `JAX` [@jax], it supports just-in-time (JIT) compilation for hardware acceleration on CPU, GPU, and TPU machines. It also supports automatic differentiation for gradient-based model learning. While other libraries exist for state space modeling in Python [@pyhsmm; @ssm; @eeasensors; @seabold2010statsmodels; @hmmlearn] and Julia [@dalle2024hiddenmarkovmodels],  `Dynamax` provides a diverse combination of low-level inference algorithms and high-level modeling objects that can support a wide range of research applications in JAX. Additionally, `Dynamax` implements parallel message passing algorithms that leverage the associative (i.e., parallel) scan primitive in JAX to take greater advantage of modern hardware accelerators. Currently, these primitives are not natively supported in other frameworks like PyTorch. While subsets of these models and algorithms in Dynamax may be found in other libraries, Dynamax is a "one stop shop" for state space modeling in JAX.
 
 The API for `Dynamax` is divided into two parts: a set of core, functionally pure, low-level inference algorithms, and a high-level, object oriented module for constructing and fitting probabilistic SSMs. The low-level inference API provides message passing algorithms for several common types of SSMs. For example, `Dynamax` provides `JAX` implementations for:
 
