+++ paper/paper.bib
@@ -222,3 +222,13 @@ year={2022}
   url={https://github.com/probml/sts-jax},
   year={2022},
   }
+
+@article{dalle2024hiddenmarkovmodels,
+  title={{HiddenMarkovModels.jl: Generic, fast and reliable state space modeling}},
+  author={Dalle, Guillaume},
+  journal={Journal of Open Source Software},
+  volume={9},
+  number={96},
+  pages={6436},
+  year={2024}
+}
