### 2025-03-14 dc9544e
slight tweaks to wording. removing hyperlink
