### 2025-03-14 probml/dynamax#386 @slinderman
Thank you for these comments @gdalle. I've updated the paper to address these questions, and I've provided further responses and clarifications below.

>> Dynamax supports canonical SSMs and allows the user to construct bespoke models as needed.

> Can you give more details on how this is implemented API-wise? For instance, how generic can observation distributions be?
> This question was a major motivation for my own take on HMMs, coded in Julia (https://joss.theoj.org/papers/10.21105/joss.06436). Hopefully I didn't misrepresent dynamax in the state of the art there.

Dynamax is quite flexible in this regard! I created a new tutorial on [_Creating Custom HMMs_](https://probml.github.io/dynamax/notebooks/hmm/custom_hmm.html) to show how it's done. Basically, you have to specify the learnable parameters (along with bijectors that map them to an unconstrained parameterization) and implement a function that returns the conditional distribution of the emissions given the latent state, $p(x_t \mid z_t)$. This function can optionally take exogenous inputs, $u_t$, too. Then the custom model inherits the base model's functions for state inference and parameter estimation. By default, the base class uses Adam (or your preferred gradient-based optimizer) for the M-step of EM, but you can also implement a closed form M-step with a bit more work.

I see that you opened an earlier [issue](https://github.com/probml/dynamax/issues/359) about the relatively slowness of Dynamax for model fitting (as opposed to state inference). It's worth double checking that the long runtimes are not due to JIT compilation. If it's not compilation related, then I suspect that the gap is because the Dynamax models use Tensorflow Probability to evaluate log probs, and that probably introduces a bunch of overhead. Someday (maybe Dynamax 2.0!) I'd like to speed that up!

>> Dynamax provides a unique combination of low-level inference algorithms and high-level modeling objects that can support a wide range of research applications in JAX.

> Would it be possible to provide concrete examples of what is missing from the previous libraries? Obviously JAX support is a big aspect, since I know that some of them are coded in Numpy (hmmlearn) or PyTorch (pomegranate)

The biggest distinctions are:
1. Since Dynamax is written in JAX, it inherits the JAX support for various backends (CPU, GPU, and TPU), whereas other frameworks would require more work to take advantage of these hardware accelerators (to the best of my knowledge).
2. Another nice feature is that Dynamax has a lot of useful inference algorithms for various types of models, including HMMs, linear Gaussian SSMs, and nonlinear/non-Gaussian SSMs. I don't know of another library that has the same breadth.
3. We also have parallel message passing routines that can run in sublinear time on GPUs using the JAX `associative_scan` primitive, as discussed below. This is a pretty unique feature compared to standard libraries!

But at the same time, I would agree that all of these models/algorithms can be found in some shape or form in other libraries, so it's really the "one stop shop" aspect of Dynamax that makes it appealing.

>> Parallel message passing routines that leverage GPU or TPU acceleration to perform message passing in sublinear time.

> What do you mean by sublinear time? Isn't it just (roughly speaking) total sequential time divided by amount of parallelism?

It's very different and super cool! The parallel filter/smoothers use an associative scan rather than a sequential scan. So rather than taking $O(T)$ time, they take only $O(\log T)$ time on a parallel machine with $O(T)$ processors. this is parallelizing over time rather than parallelizing over batches! Our implementation takes advantage of the `associative_scan` primitive in JAX, which is not (yet) natively support in PyTorch. We've made use of these primitives in our own research projects, like deep state space modeling with [S5](https://arxiv.org/abs/2208.04933).

There are some caveats to keep in mind: even on a GPU, you can exhaust the parallel resources, so you see diminishing returns for large enough $T$. These algorithms also require more memory usage and compute per step (it involves matrix-matrix multiplication rather than matrix-vector). Altogether, the final wall-clock time ends up depending a lot on your hardware.

>> While other libraries exist for state space modeling in Python [...], Dynamax provides a unique combination of low-level inference algorithms and high-level modeling objects that can support a wide range of research applications in JAX.

> I would like to see a more detailed discussion of the state of the field: what do you think is missing from existing packages? Is it just that some of them are not written in JAX? Does writing in JAX allow you to do things you aren't able to achieve with e.g. PyTorch?

I've updated the paper to include a longer discussion of the distinguishing features of Dynamax. As I said above though, various subsets of the models and algorithms in Dynamax can be found in other packages, and I wouldn't say that there's something that exists in Dynamax that can't be found anywhere else. It's really the combination of models and algorithms in JAX that makes Dynamax special. Hopefully the paper is more clear on this now.

Thanks again for your helpful suggestions!
