### 2025-03-17 8dba507
Bugfix InverseWishart scale matrix
