### 2025-03-20 probml/dynamax#386 @gdalle
Thanks for these updates! I actually learned in the meantime what a parallel scan is, and this makes so much more sense :)
The paper looks good to me in its current state
