### 2025-03-20 probml/dynamax#377 @gdalle
Sounds good to me!
