### 2025-03-20 probml/dynamax#376 @gdalle
Thank you for addressing my remarks!

> I'm not sure what exactly this would look like. The README and the home page of the documentation provides a basic map, and the demo notebooks are organized into categories for the major types of models and algorithms we support. Perhaps you could point me to an example of what a more explanatory page would look like? Divio points to [this page](https://docs.divio.com/background/) as an example, but that feels a bit redundant to me. I'm still open to this suggestion though.

Okay, that's fair, I guess an organization by HMM type works fine.

> It's certainly more expensive than doing least squares once with the full batch of data since you are updating the filtering distribution with each new data point, and that is cubic in the number of parameters. So for online linear regression, the cost is O ( N D 2 + T D 3 ) where D is the parameter dimension. It is more efficient than naively running least squares T times though, since it recursively updates the posterior with each new data point. I.e., as a function of the number of data points N , the cost is only O ( N D 2 ) rather than O ( T N D 2 ) if you naively ran least squares T times. Not sure it's worth spelling these details out in the notebook, but we can if you think it's important.

No, I don't think it is necessary indeed.

> We've also fixed the plots adjusted the HMC hyperparameters
so that the curves look more reasonable.

When you freeze a subset of the parameters, isn't it weird that the HMC loglikelihood curve _starts_ above the likelihood of the true params?

Overall the documentation is now fully satisfactory, closing this issue now.
