### 2025-03-24 c338c5c
Bugfix InverseWishart variance calculation
