### 2025-03-24 probml/dynamax#410 @osorensen
Editorial comment on JOSS submission
This is an editorial comment on your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/7069):

Please add the correct DOI `10.21105/joss.06436` to this reference:

https://github.com/probml/dynamax/blob/0a7c01f41cbff9d9124be0b2124cdb3a045f76aa/paper/paper.bib#L226-L234
