### 2025-03-24 #7069 @osorensen
@slinderman, I have now read through the paper once more, and my only suggested edit is that you add the missing DOI mentioned in [this issue](https://github.com/probml/dynamax/issues/410).
