### 2025-03-24 e8d37be
adding doi for dalle 2024
