+++ paper/paper.bib
@@ -230,5 +230,6 @@ year={2022}
   volume={9},
   number={96},
   pages={6436},
-  year={2024}
+  year={2024},
+  doi={10.21105/joss.06436}
 }
