### 2025-03-24 probml/dynamax#410 @slinderman
Thank you @osorensen. I added the DOI is [this commit](https://github.com/probml/dynamax/commit/e8d37beb97bbc518c18cace5709f9c574c965af4).
