### 2025-03-24 #7069 @slinderman
Thank you @thomaspinder for your review, and thank you @osorensen for catching the DOI. I've added the missing reference and closed the issue.
