### 2025-03-25 #7069 @slinderman
Thank you @osorensen. I have released Dynamax-1.0.0 to PyPI and uploaded a copy of the source code to figshare [here](https://figshare.com/account/articles/28665131?file=53231987). The DOI is https://doi.org/10.6084/m9.figshare.28665131
