### 2025-03-26 #7069 @osorensen
Thanks @slinderman. To your question, you should merge the `paper` branch into the `main` branch so that the version accepted by JOSS corresponds to release 1.0.0 on `main`.
