### 2025-03-26 #7069 @osorensen
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/14077232961?check_suite_focus=true). The generated XML metadata file is invalid.
>
> Element isbn: [facet 'minLength'] The value has a length of '9'; this underruns the allowed minimum length of '10'.

@slinderman, could you try deleting the following line? It's my best guess at what causes the issue above.

https://github.com/probml/dynamax/blob/83b8cca769b3877d91283a56520d8b3479f6c36a/paper/paper.bib#L78

Please report here when done, and I'll try running the paper acceptance again.
