### 2025-03-26 7ecd3d7
removing isbn from weinreb et al
