+++ paper/paper.bib
@@ -75,7 +75,6 @@
   author = {Weinreb, Caleb and Pearl, Jonah E. and Lin, Sherry and Osman, Mohammed Abdal Monium and Zhang, Libby and Annapragada, Sidharth and Conlin, Eli and Hoffmann, Red and Makowska, Sofia and Gillis, Winthrop F. and Jay, Maya and Ye, Shaokai and Mathis, Alexander and Mathis, Mackenzie W. and Pereira, Talmo and Linderman, Scott W. and Datta, Sandeep Robert},
   date = {2024/07/01},
   id = {Weinreb2024},
-  isbn = {1548-7105},
   journal = {Nature Methods},
   number = {7},
   pages = {1329--1339},
