### 2025-03-26 #7069 @slinderman
Ok, I deleted the isbn field from that reference and merged the paper branch into main. (See above.)
