### 2025-03-26 #7069 @slinderman
Looks good to me! On Mar 26, 2025, at 7:56 AM, The Open Journals editorial robot ***@***.***> wrote:﻿
  👋 @openjournals/dsais-eics, this paper is ready to be accepted and published.
Check final proof 👉📄 Download article
If the paper PDF and the deposit XML files look good in openjournals/joss-papers#6537, then you can now move forward with accepting the submission by compiling again with the command @editorialbot accept—Reply to this email directly, view it on GitHub, or unsubscribe.You are receiving this because you were mentioned.Message ID: ***@***.***>

editorialbot left a comment (openjournals/joss-reviews#7069)
👋 @openjournals/dsais-eics, this paper is ready to be accepted and published.
Check final proof 👉📄 Download article
If the paper PDF and the deposit XML files look good in openjournals/joss-papers#6537, then you can now move forward with accepting the submission by compiling again with the command @editorialbot accept
