### 2025-04-03 #7069 @arfon
@slinderman - I'm helping out @crvernon with some of his papers. Could you please merge this PR which has some tiny formatting fixes: https://github.com/probml/dynamax/pull/413
