### 2025-04-03 #7069 @slinderman
It looks good to me!Thank you,ScottOn Apr 3, 2025, at 7:23 AM, The Open Journals editorial robot ***@***.***> wrote:﻿
  👋 @openjournals/dsais-eics, this paper is ready to be accepted and published.
Check final proof 👉📄 Download article
If the paper PDF and the deposit XML files look good in openjournals/joss-papers#6578, then you can now move forward with accepting the submission by compiling again with the command @editorialbot accept—Reply to this email directly, view it on GitHub, or unsubscribe.You are receiving this because you were mentioned.Message ID: ***@***.***>

editorialbot left a comment (openjournals/joss-reviews#7069)
👋 @openjournals/dsais-eics, this paper is ready to be accepted and published.
Check final proof 👉📄 Download article
If the paper PDF and the deposit XML files look good in openjournals/joss-papers#6578, then you can now move forward with accepting the submission by compiling again with the command @editorialbot accept
